% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{add_chip_peak_labels_to_sites}
\alias{add_chip_peak_labels_to_sites}
\title{Add ChIP-seq peak labels to the candidate sites}
\usage{
add_chip_peak_labels_to_sites(
  sites.df,
  chip_peak_file = NULL,
  chip_peak_sampleID = NULL,
  chip_peak_dir = "./"
)
}
\arguments{
\item{sites.df}{data frame containing the candidate sites}

\item{chip_peak_file}{ChIP-seq peak file (.bed.gz format)}

\item{chip_peak_sampleID}{ENCODE sample ID of ChIP-seq peaks (.bed.gz format)}

\item{chip_peak_dir}{Directory of ChIP-seq peaks}
}
\description{
Add ChIP-seq peak labels to the candidate sites
}
