% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numbers.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers}
\usage{
add_numbers(a, b)
}
\arguments{
\item{a}{numeric number}

\item{b}{numeric number}
}
\value{
numeric
}
\description{
Add two numeric numbers.
}
\examples{
add_numbers(10, 5)
}
