% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_training_data.R
\name{assemble_partition_training_data}
\alias{assemble_partition_training_data}
\title{Assemble TOP training data for all TF x cell type combinations
in a selected partition}
\usage{
assemble_partition_training_data(
  tf_cell_table,
  logistic_model = FALSE,
  chip_col = "chip",
  training_chrs = paste0("chr", seq(1, 21, 2)),
  n_partitions = 10,
  part = 1,
  max_sites = 5000,
  seed = 1
)
}
\arguments{
\item{tf_cell_table}{A data frame containing the information about
the training TFs, cell types and paths of the training data.
The first three columns should contain TF name, cell type,
and file name containing the corresponding training data.}

\item{logistic_model}{Logical; if TRUE, use the logistic version of TOP model.}

\item{chip_col}{The column name of ChIP data in the combined data
(default: \dQuote{chip}).}

\item{training_chrs}{Chromosomes used for training the model
Default: odd chromosomes.}

\item{n_partitions}{Number of partitions to split the training data
(default: 10).}

\item{part}{An integer between 1 and \code{n_partitions} to specify the
partition to assemble the training data for.}

\item{max_sites}{Max number of candidate sites to keep in each partition
(default: 5000). Randomly sample \code{max_sites} sites in each partition
if the number of candidate sites in each partition exceeds \code{max_sites}.}

\item{seed}{seed used when sampling sites.}
}
\value{
A data frame of the training data
with all TF x cell type combinations in the selected partition.
}
\description{
Prepare the training data for fitting TOP models. It first loads
the training data for each of the training TF x cell type combinations from
\code{tf_cell_table}, then selects the candidate sites in training chromosomes,
and splits the selected candidate sites into partitions (default: 10), and
assembles training data for all TF x cell type combinations for a selected partition.
}
