% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dnase_data.R
\name{bam_sort_index_stats}
\alias{bam_sort_index_stats}
\title{Sort and index the BAM file, and then retrieve and print stats in the index file.}
\usage{
bam_sort_index_stats(
  bam_file,
  dir_output = NA,
  sort = TRUE,
  stats = TRUE,
  path_samtools = "samtools"
)
}
\arguments{
\item{bam_file}{Input BAM file.}

\item{dir_output}{Output directory.}

\item{sort}{logical. If TRUE, sort (and index) the BAM file.}

\item{stats}{logical. If TRUE, retrieve and print stats in the index file.}

\item{path_samtools}{Path to samtools executable.}
}
\description{
Sort and index the BAM file, and then retrieve and print stats in the index file.
}
