% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{bin_dnase}
\alias{bin_dnase}
\title{Binning and transform DNase count matrix}
\usage{
bin_dnase(
  dnase_counts,
  bin.method = c("M5", "M1", "M2", "M3", "M12", "M24"),
  transform = c("asinh", "log2")
)
}
\arguments{
\item{dnase_counts}{DNase (or ATAC) count matrix}

\item{bin.method}{MILLIPEDE binning scheme (Default: 'M5').}

\item{transform}{asinh or log2 transform of DNase counts}
}
\description{
Binning DNase using MILLIPEDE binning and
then take asinh (or log2) transform
}
