% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{combine_dnase_bins_chip_data}
\alias{combine_dnase_bins_chip_data}
\title{Combined normalized and binned DNase data with normalized ChIP counts data}
\usage{
combine_dnase_bins_chip_data(
  dnase_counts_file,
  chip_counts_file,
  bin.method = "M5"
)
}
\arguments{
\item{dnase_counts_file}{Normalized DNase counts files}

\item{chip_counts_file}{Normalized ChIP counts files}

\item{bin.method}{MILLIPEDE binning scheme (Default: 'M5')}
}
\description{
Combined normalized and binned DNase data with
normalized ChIP counts data. This function takes normalized
DNase and ChIP data with replicates merged.
}
