% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_utils.R
\name{combine_dnase_counts}
\alias{combine_dnase_counts}
\title{Combine DNase tagcount matrices in both forward and reverse strands}
\usage{
combine_dnase_counts(file_dnase_matrix_fwd, file_dnase_matrix_rev)
}
\arguments{
\item{file_dnase_matrix_fwd}{filename for DNase tagcount matrix in forward strand}

\item{file_dnase_matrix_rev}{filename for DNase tagcount matrix in reverse strand}
}
\description{
Combine DNase tagcount matrices in both forward and reverse strands
}
