% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_model_utils.R
\name{combine_TOP_samples}
\alias{combine_TOP_samples}
\title{Combine and take the average of TOP posterior samples from all partitions}
\usage{
combine_TOP_samples(TOP_samples_files, thin = 1, n.samples = 1000)
}
\arguments{
\item{TOP_samples_files}{file names of the posterior samples from all partitions}

\item{thin}{thinning rate of extract the posterior samples,
must be a positive integer.}

\item{n.samples}{Randomly choose n.samples posterior samples,
if the number of posterior samples is greater than n.samples}
}
\value{

}
\description{
Combine and take the average of TOP posterior samples from all partitions
}
