% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TOP_model_jags.R
\name{combine_training_data}
\alias{combine_training_data}
\title{Combine training data from TF x cell type combos}
\usage{
combine_training_data(tf_cell_table, total_partition = 10, n_part = 1)
}
\arguments{
\item{tf_cell_table}{a data frame with TF name, PWM name, cell type, and link to the training data with PWM, DNase and ChIP data.}

\item{total_partition}{split the data into partitions (default = 10) and run Gibbs sampling in parallel.}

\item{n_part}{which partition to use for training the model}
}
\value{
The function returns a data frame of training data with all TFs and cell type combos.
}
\description{
Combine training data from TF x cell type combos
}
