% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteCasesScoreTestSupport.R
\name{CompleteCasesScoreTestSupport}
\alias{CompleteCasesScoreTestSupport}
\title{Title}
\usage{
CompleteCasesScoreTestSupport(y, baselineonly, additive, pairwise.interaction,
  saturated)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
