% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{compute_mapability}
\alias{compute_mapability}
\title{Title}
\usage{
compute_mapability(
  sites,
  mapability_file,
  bigWigAverageOverBed_path = "bigWigAverageOverBed"
)
}
\arguments{
\item{sites}{A data frame of candidate binding sites with the first
6 columns the same as in the BED format}

\item{mapability_file}{ENCODE mapability bigWig file}

\item{bigWigAverageOverBed_path}{path of bigWigAverageOverBed executable}
}
\description{
Title
}
