% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_ATAC_genome_cuts_readpairs}
\alias{count_ATAC_genome_cuts_readpairs}
\title{Count genome coverage of shifted ATAC 5' cuts for ATAC-seq paired-end reads}
\usage{
count_ATAC_genome_cuts_readpairs(
  bam_file,
  select_NFR_fragments = FALSE,
  shift_ATAC = FALSE,
  write_bedgraph = FALSE,
  outdir = dirname(bam_file),
  outname
)
}
\arguments{
\item{bam_file}{BAM file of ATAC-seq paired-end reads}

\item{outdir}{Output directory. Default: the directory of the \code{bam_file}}

\item{plot_fragsize}{Logical, if TRUE, plot the distribution of fragment sizes}
}
\description{
Count genome coverage of shifted ATAC 5' cuts for ATAC-seq paired-end reads
}
