% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{count_dnase_genome_cuts}
\alias{count_dnase_genome_cuts}
\title{Count DNase-seq or ATAC-seq cleavage for a given genome.}
\usage{
count_dnase_genome_cuts(
  bam_file,
  chrom_size_file,
  outdir,
  outname,
  bedtools_path = "bedtools",
  bedGraphToBigWig_path = "bedGraphToBigWig",
  bedSort_path = "bedSort"
)
}
\arguments{
\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File name of genome sizes by chromosomes.}

\item{outdir}{Output directory.}

\item{outname}{Output prefix.}

\item{bedtools_path}{Path to bedtools executable.}

\item{bedGraphToBigWig_path}{Path to UCSC bedGraphToBigWig executable.}

\item{bedSort_path}{Path to UCSC bedSort executable.}
}
\description{
Count DNase-seq or ATAC-seq cleavage for a given genome.
}
