% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dnase_data.R
\name{count_genome_coverage}
\alias{count_genome_coverage}
\title{Count the coverage of DNase-seq cuts for a given genome.}
\usage{
count_genome_coverage(
  bam_file,
  file_out = NA,
  strand = "+",
  file_genome_sizes = NA,
  path_samtools = "samtools",
  path_bedtools = "bedtools",
  path_bedGraphToBigWig = "bedGraphToBigWig",
  output_format = "Bigwig"
)
}
\arguments{
\item{bam_file}{Input BAM file.}

\item{file_out}{Output file name.}

\item{strand}{strand: '+' or '-'.}

\item{file_genome_sizes}{File name of genome sizes by chromosomes.}

\item{path_samtools}{Path to samtools executable.}

\item{path_bedtools}{Path to bedtools executable.}

\item{path_bedGraphToBigWig}{Path to bedGraphToBigWig executable.}

\item{output_format}{Output format: 'Bigwig' or 'bedGraph'.}
}
\description{
Count the coverage of DNase-seq cuts for a given genome.
}
