% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_genome_cuts}
\alias{count_genome_cuts}
\title{Count DNase-seq or ATAC-seq cleavage along the genome}
\usage{
count_genome_cuts(
  bam_file,
  chrom_size_file,
  outdir,
  outname,
  bedtools_path = "bedtools",
  bedGraphToBigWig_path = "bedGraphToBigWig",
  bedSort_path = "bedSort"
)
}
\arguments{
\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File of genome sizes by chromosomes.}

\item{outdir}{Output directory (default: use the directory of bam_file).}

\item{outname}{Output prefix (default: use the prefix of bam_file).}

\item{bedtools_path}{Path to bedtools executable.}

\item{bedGraphToBigWig_path}{Path to UCSC bedGraphToBigWig executable.}

\item{bedSort_path}{Path to UCSC bedSort executable.}
}
\description{
Count genomic cleavage (5' end) from DNase or ATAC-seq bam files
using the \code{genomecov} tool from \code{bedtools}.
Then save the output in Bigwig format using
the \code{bedGraphToBigWig} tool
(and \code{bedSort} if available) from UCSC.
}
