% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_genome_cuts_ATACreadpairs}
\alias{count_genome_cuts_ATACreadpairs}
\title{Count genome coverage of shifted ATAC 5' cuts for ATAC-seq paired-end reads}
\usage{
count_genome_cuts_ATACreadpairs(
  bam_file,
  select_NFR_fragments = FALSE,
  shift_ATAC = FALSE,
  outdir = dirname(bam_file),
  outname,
  bedGraphToBigWig_path = "bedGraphToBigWig"
)
}
\arguments{
\item{bam_file}{BAM file of ATAC-seq paired-end reads}

\item{select_NFR_fragments}{Select NRF size fragments}

\item{shift_ATAC}{If TRUE, shift ATAC-seq reads}

\item{outdir}{Output directory (default: use the directory of \code{bam_file}).}

\item{outname}{Output prefix (default: use the prefix of \code{bam_file}).}

\item{bedGraphToBigWig_path}{Path to UCSC \code{bedGraphToBigWig} executable.}
}
\description{
Count genome coverage of shifted ATAC 5' cuts for ATAC-seq paired-end reads
}
