% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_genome_cuts_nobedtools}
\alias{count_genome_cuts_nobedtools}
\title{Count DNase-seq or ATAC-seq cuts along the genome (without using bedtools)}
\usage{
count_genome_cuts_nobedtools(
  bam_file,
  chrom_size_file,
  shift_ATAC = FALSE,
  shift_ATAC_bases = c(4L, -4L),
  outdir = dirname(bam_file),
  outname,
  bedGraphToBigWig_path = "bedGraphToBigWig"
)
}
\arguments{
\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File of genome sizes by chromosomes.}

\item{shift_ATAC}{Logical. When \code{shift_ATAC = TRUE},
it shifts reads according to \code{shift_ATAC_bases}.}

\item{shift_ATAC_bases}{Number of bases to shift.}

\item{outdir}{Output directory (default: use the directory of \code{bam_file}).}

\item{outname}{Output prefix (default: use the prefix of \code{bam_file}).}

\item{bedGraphToBigWig_path}{Path to UCSC \code{bedGraphToBigWig} executable.}
}
\description{
This is an alternative function to the count_genome_cuts(),
but without using \code{bedtools}.
This may be slower than count_genome_cuts() for large BAM files.
}
\examples{
count_genome_cuts_nobedtools(bam_file='K562.ATAC.bam',
                             chrom_size_file='hg38.chrom.sizes',
                             shift_ATAC=TRUE,
                             outdir='processed_data',
                             outname='K562.ATAC.bam',
                             bedGraphToBigWig_path='bedGraphToBigWig')
}
