% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{count_normalize_chip}
\alias{count_normalize_chip}
\title{Count and normalize ChIP-seq read coverage}
\usage{
count_normalize_chip(
  pwm_id,
  cell_type,
  training_metadata,
  sites_file,
  chrom_size_file,
  chip_dir,
  outdir,
  outname,
  ref.size = 1e+07,
  bedtools_path = "bedtools"
)
}
\arguments{
\item{pwm_id}{Motif pwm ID}

\item{cell_type}{Cell type}

\item{training_metadata}{training metadata}

\item{sites_file}{Filename for candidate sites}

\item{chrom_size_file}{Filename for chromosome sizes}

\item{chip_dir}{Directory of ChIP-seq bam files}

\item{outdir}{Output directory}

\item{outname}{Output filename prefix}

\item{ref.size}{ChIP-Seq reference library size (Default: 10 million)}

\item{bedtools_path}{Path to bedtools executable}
}
\description{
Count and normalize ChIP-seq read coverage
}
