% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountTumorNumber.R
\name{CountCovarNumber}
\alias{CountCovarNumber}
\title{Count variable number}
\usage{
CountCovarNumber(covar)
}
\arguments{
\item{covar}{the covariate matrix}
}
\description{
Count variable number
}
\keyword{internal}
