% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{CPOP_lambdaPlot}
\alias{CPOP_lambdaPlot}
\title{Title}
\usage{
CPOP_lambdaPlot(
  CPOP_model,
  nFeatures = 20,
  s = "lambda.min",
  interactive = FALSE,
  label = FALSE
)
}
\arguments{
\item{CPOP_model}{A CPOP model}

\item{nFeatures}{The number of features to plot, features are ranked beta's for lambda.min}

\item{s}{lambda value. Default is "lambda.min"}
}
\description{
Title
}
\examples{
data(cpop_data_binary, package = "CPOP")

x1 <- cpop_data_binary$x1
x2 <- cpop_data_binary$x2
x3 <- cpop_data_binary$x3
y1 <- cpop_data_binary$y1
y2 <- cpop_data_binary$y2
y3 <- cpop_data_binary$y3

set.seed(23)
x_list <- list(x1, x2)
y_list <- list(factor(y1), factor(y2))

fCPOP_model <- Frankenstein_CPOP(x_list, y_list)
CPOP_coefPlot(fCPOP_model)

}
