% --- Source file: data.Rd ---
\name{data}
\alias{data}
%\docType{package}
\title{
 Example data 
}
\description{
 Sample covariate, outcome and SNP data data 
}
\details{ This is simulated breast cancer data used in the examples.
}
\examples{
 # Load and print the first 5 rows 
 data(data, package="TOP") 

 data[1:5, ]
}
\keyword{data}




