% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_95CI_p.R
\name{DisplayFirstStageTestResult}
\alias{DisplayFirstStageTestResult}
\title{Display the first stage parameters}
\usage{
DisplayFirstStageTestResult(logodds, sigma, places = 5)
}
\arguments{
\item{logodds}{the log odds ratio vectors}

\item{sigma}{the covariance matrix for the log odds ratio vectors}

\item{places}{the numerical places to keep}
}
\description{
Display the first stage parameters
}
