% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_95CI_p.R
\name{DisplayMixedScoreTestResult}
\alias{DisplayMixedScoreTestResult}
\title{Calculate the mixed effect model p value based on score and information matrix}
\usage{
DisplayMixedScoreTestResult(score.fix, infor.fix, score.random, infor.random)
}
\arguments{
\item{score.fix}{fixed effect score}

\item{infor.fix}{fixed effect information matrix}

\item{score.random}{random effect score}

\item{infor.random}{random effect information matrix}
}
\description{
Calculate the mixed effect model p value based on score and information matrix
}
