% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmvpoly.R
\name{EMmvpoly}
\alias{EMmvpoly}
\title{Two-stage model MLE estimation with EM algorithm}
\usage{
EMmvpoly(y, baselineonly = NULL, additive = NULL,
  pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888, delta0 = NULL, cutoff = 10)
}
\arguments{
\item{y}{the phenotype file. The first column is the case control disease status. The other columns are the tumor characteristics status}

\item{baselineonly}{the covariates to be adjusted used baseline effect only model. This assumes the odds ratio of the covariates for all the subtpes to be the same.}

\item{additive}{the covariates to be adjusted used the additive two-stage model}

\item{pairwise.interaction}{the covariates to be adjusted used the pairwise interaction two-stage model}

\item{saturated}{the covariates to be adjusted used the saturated two-stage model. This model assumes every subtype has their specific odds ratio. It's equivalent to the polytmous model.}

\item{missingTumorIndicator}{The indicators to show the tumor characteristics are missing. In the example, we put missing tumor characteristics as 888. Note, for all the controls subjects, they don't have tumor characteristics. So their tumor characteristics are put as NA instead of 888 to differentiate with cases missing tumor characteristics.}

\item{delta0}{the starting value for the second stage parameters. By defualt, we will use the empirical distribution of the subtypes.}

\item{cutoff}{by default, the model will remove the subtypes with less than 10 cases, the user can specify other values by changing the cutoff. But we don't recommend to set the cutoff too low, since the asymptotic convergence requires enough sample size}
}
\description{
Two-stage model MLE estimation with EM algorithm
}
\keyword{internal}
