% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMScoreTestSupport.R
\name{EMScoreTestSupport}
\alias{EMScoreTestSupport}
\title{Title}
\usage{
EMScoreTestSupport(
  y,
  baselineonly,
  additive,
  pairwise.interaction,
  saturated,
  missingTumorIndicator,
  cutoff = 10
)
}
\arguments{
\item{cutoff}{}
}
\description{
Title
}
