% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMStepScoreTestMixedModel.R,
%   R/EMStepScoreTestSupportMixedModel.R
\name{EMStepScoreTestSupportMixedModel}
\alias{EMStepScoreTestSupportMixedModel}
\alias{EMStepScoreTestSupportMixedModel}
\title{Title}
\usage{
EMStepScoreTestSupportMixedModel(delta0, y, x.all, z.standard, z.all,
  missingTumorIndicator)

EMStepScoreTestSupportMixedModel(delta0, y, x.all, z.standard, z.all,
  missingTumorIndicator)
}
\arguments{
\item{missingTumorIndicator}{}

\item{missingTumorIndicator}{}
}
\description{
Title

Title
}
