% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_model_utils.R
\name{extract_tf_cell_combos}
\alias{extract_tf_cell_combos}
\title{Extract a table listing the indices and names of TF and cell type
combinations from the assembled training data}
\usage{
extract_tf_cell_combos(assembled_training_data)
}
\arguments{
\item{assembled_training_data}{Assembled training data as in the
\code{assemble_training_data} function}
}
\value{
a data frame of the indices and names of TF x cell type combinations.
}
\description{
Extract a table listing the indices and names of TF x cell type
combinations from the assembled training data.
This will be used to extract regression coefficients from the TOP fit result.
}
