% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_model_utils.R
\name{extract_TOP_coef_samples}
\alias{extract_TOP_coef_samples}
\title{Extract alpha and beta coefficients from TOP posterior samples}
\usage{
extract_TOP_coef_samples(
  TOP_samples,
  tf_cell_combos,
  tf_name,
  cell_type,
  n.bins = 5,
  level = c("bottom", "middle", "top")
)
}
\arguments{
\item{TOP_samples}{TOP samples combined from all partitions.}

\item{tf_cell_combos}{a table of TF x cell type combinations,
included in the training data.}

\item{tf_name}{TF name of interest.}

\item{cell_type}{Cell type of interest.}

\item{n.bins}{Number of DNase or ATAC bins in TOP model (default = 5).}

\item{level}{The level in the TOP model (options: bottom, middle, or top),
'bottom' level: TF- and cell-type- specific,
'middle' level: TF-specific, cell-type generic,
'top' level: TF-generic}
}
\description{
Extract alpha and beta coefficients from TOP posterior samples
}
