% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{filter_blacklist}
\alias{filter_blacklist}
\title{Filter candidate sites in ENCODE blacklist regions}
\usage{
filter_blacklist(sites.df, blacklist_file)
}
\arguments{
\item{sites.df}{A data frame of candidate binding sites}

\item{blacklist_file}{ENCODE blacklist file}
}
\description{
Filter candidate sites in ENCODE blacklist regions
}
