% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TOP_model_jags.R
\name{fit_TOP_jags}
\alias{fit_TOP_jags}
\title{Fit TOP model using JAGS with R2jags package}
\usage{
fit_TOP_jags(
  data.train,
  TOP.model,
  parameters.to.save,
  n.iter = 1e+05,
  n.burnin = 5000,
  n.thin = 2,
  n.chains = 1,
  DIC = TRUE
)
}
\arguments{
\item{data.train}{combined training data.}

\item{TOP.model}{TOP model written in BUGS code.}

\item{parameters.to.save}{character vector of the names of the parameters to
save which should be monitored.}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.}

\item{n.thin}{thinning rate, must be a positive integer.}

\item{n.chains}{number of Markov chains.}

\item{DIC}{logical; if TRUE (default), compute deviance, pD, and DIC.}
}
\value{

}
\description{
Fit TOP model using JAGS with R2jags package
}
