% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TOP_models.R
\name{fit_TOP_logistic_M5_model}
\alias{fit_TOP_logistic_M5_model}
\title{Fit TOP logistic model}
\usage{
fit_TOP_logistic_M5_model(
  data,
  model,
  n.iter = 10000,
  n.burnin = 5000,
  n.chains = 3,
  n.thin = 10
)
}
\arguments{
\item{data}{combined training data.}

\item{model}{TOP logistic model (written in BUGS code).}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.}

\item{n.chains}{number of Markov chains.}

\item{n.thin}{thinning rate, must be a positive integer.}
}
\description{
Fit TOP logistic model
}
