% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TOP_models.R
\name{fit_TOP_M5_model_jags}
\alias{fit_TOP_M5_model_jags}
\title{Fit TOP model using JAGS}
\usage{
fit_TOP_M5_model_jags(
  data,
  model.file,
  transform = c("asinh", "log2", "sqrt", "none"),
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  n.chains = 3,
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  quiet = FALSE
)
}
\arguments{
\item{data}{combined training data.}

\item{model.file}{TOP model (written in BUGS code).}

\item{transform}{Transformation of ChIP counts (asinh, log2, sqrt, or none).}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.}

\item{n.chains}{number of Markov chains.}

\item{n.thin}{thinning rate, must be a positive integer.}

\item{quiet}{Logical, whether to suppress stdout.}
}
\description{
Fit TOP model using JAGS
}
