% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frankenstein_CPOP.R
\name{Frankenstein_CPOP}
\alias{Frankenstein_CPOP}
\title{Run Frankenstein CPOP}
\usage{
Frankenstein_CPOP(
  x_list,
  y_list,
  covariates = NULL,
  dataset_weights = NULL,
  sample_weights = FALSE,
  optimiseExponent = FALSE,
  nCores = 1
)
}
\arguments{
\item{x_list}{PARAM_DESCRIPTION}

\item{y_list}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: NULL}

\item{dataset_weights}{PARAM_DESCRIPTION, Default: NULL}

\item{sample_weights}{PARAM_DESCRIPTION, Default: FALSE}

\item{optimiseExponent}{PARAM_DESCRIPTION, Default: FALSE}

\item{nCores}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
