% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCompleteYPheno.R
\name{GenerateCompleteXCovariates}
\alias{GenerateCompleteXCovariates}
\title{Title}
\usage{
GenerateCompleteXCovariates(y.pheno, x.all, missingTumorIndicator)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
