% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCompleteYPheno.R
\name{GenerateCompleteYPheno}
\alias{GenerateCompleteYPheno}
\title{Title}
\usage{
GenerateCompleteYPheno(y.pheno, missingTumorIndicator)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
