% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCovarName.R
\name{GenerateCovarName}
\alias{GenerateCovarName}
\title{Title}
\usage{
GenerateCovarName(baselineonly, additive, pairwise.interaction, saturated,
  x.self.design = NULL)
}
\arguments{
\item{x.self.design}{}
}
\description{
Title
}
