% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFirstStageMat.R
\name{GenerateFirstStageMat}
\alias{GenerateFirstStageMat}
\title{Title}
\usage{
GenerateFirstStageMat(beta, covar.names, subtypes.names)
}
\arguments{
\item{subtypes.names}{}
}
\description{
Title
}
