% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCompleteYPheno.R
\name{GenerateMissingPosition}
\alias{GenerateMissingPosition}
\title{Title}
\usage{
GenerateMissingPosition(y.pheno, missingTumorIndicator)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
