% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSelfCovarName.R
\name{GenerateSelfCovarName}
\alias{GenerateSelfCovarName}
\title{Title}
\usage{
GenerateSelfCovarName(x.self.design, baselineonly, additive,
  pairwise.interaction, saturated)
}
\arguments{
\item{saturated}{}
}
\description{
Title
}
