% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSelfFirstStageMat.R
\name{GenerateSelfFirstStageMat}
\alias{GenerateSelfFirstStageMat}
\title{Title}
\usage{
GenerateSelfFirstStageMat(beta, x.self.design, z.design, covar.names,
  subtypes.names)
}
\arguments{
\item{subtypes.names}{}
}
\description{
Title
}
