% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSelfXALL.R
\name{GenerateSelfXAll}
\alias{GenerateSelfXAll}
\title{Title}
\usage{
GenerateSelfXAll(y, x.self.design, baselineonly, additive, pairwise.interaction,
  saturated)
}
\arguments{
\item{saturated}{}
}
\description{
Title
}
