% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTumorCharacterCat.R
\name{GenerateTumorCharacterCat}
\alias{GenerateTumorCharacterCat}
\title{Title}
\usage{
GenerateTumorCharacterCat(y.pheno.complete)
}
\arguments{
\item{y.pheno.complete}{}
}
\description{
Title
}
