% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_z_design_matrix.R
\name{GenerateZDesignAdditive}
\alias{GenerateZDesignAdditive}
\title{Title}
\usage{
GenerateZDesignAdditive(tumor.character.cat, tumor.number, tumor.names,
  freq.subtypes)
}
\arguments{
\item{freq.subtypes}{}
}
\description{
Title
}
