% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateZDesignCombination.R
\name{GenerateZDesignCombination}
\alias{GenerateZDesignCombination}
\title{Generate the three Z design matrix given the tumor characteristics}
\usage{
GenerateZDesignCombination(y, missingTumorIndicator = 888, cutoff = 10)
}
\arguments{
\item{y}{the phenotype file. The first column is the case control disease status. The other columns are the tumor characteristics status}
}
\description{
Generate the three Z design matrix given the tumor characteristics
}
