% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_z_design_matrix.R
\name{GenerateZDesignNamesAdditive}
\alias{GenerateZDesignNamesAdditive}
\title{Title}
\usage{
GenerateZDesignNamesAdditive(tumor.names)
}
\arguments{
\item{tumor.names}{}
}
\description{
Title
}
