% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_z_design_matrix.R
\name{GenerateZDesignPairwiseInteraction}
\alias{GenerateZDesignPairwiseInteraction}
\title{Title}
\usage{
GenerateZDesignPairwiseInteraction(tumor.character.cat, tumor.number,
  tumor.names, freq.subtypes)
}
\arguments{
\item{freq.subtypes}{}
}
\description{
Title
}
