% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{get_dnase_sites_counts}
\alias{get_dnase_sites_counts}
\title{Get DNase count matrices for candidate sites.}
\usage{
get_dnase_sites_counts(
  sites_file,
  dnase_fwd_count_file,
  dnase_rev_count_file,
  dnase_fwd_matrix_file,
  dnase_rev_matrix_file,
  bwtool_path = "bwtool"
)
}
\arguments{
\item{sites_file}{Filename for candidate sites}

\item{dnase_fwd_count_file}{Filename for DNase counts in forward strand (Bigwig format)}

\item{dnase_rev_count_file}{Filename for DNase counts in reverse strand (Bigwig format)}

\item{dnase_fwd_matrix_file}{Filename for DNase count matrix in forward strand}

\item{dnase_rev_matrix_file}{Filename for DNase count matrix in reverse strand}

\item{bwtool_path}{Path to bwtool executable.}
}
\description{
Get DNase count matrices for candidate sites.
}
