% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{get_genome_fasta}
\alias{get_genome_fasta}
\title{Retrieve FASTA genome file}
\usage{
get_genome_fasta(
  BSgenome,
  outfile = "genome.fa",
  chr_list = paste0("chr", c(1:22, "X", "Y", "M"))
)
}
\arguments{
\item{BSgenome}{BSgenome object. For example,
\code{BSgenome.Hsapiens.UCSC.hg19}, \code{BSgenome.Hsapiens.UCSC.hg38}, etc.}

\item{outfile}{Output FASTA (\code{.fa}) file}
}
\description{
Retrieve FASTA genome file
}
