% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{get_sites_counts}
\alias{get_sites_counts}
\title{Get count matrices around candidate sites.}
\usage{
get_sites_counts(
  sites_file,
  genomecount_dir,
  genomecount_name,
  bwtool_path = "bwtool"
)
}
\arguments{
\item{sites_file}{Filename for candidate sites}

\item{genomecount_dir}{Directory for genome counts}

\item{genomecount_name}{Filename prefix for genome counts}

\item{bwtool_path}{Path to bwtool executable.}
}
\description{
Get count matrices around candidate sites.
}
