% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{get_sites_counts_slower}
\alias{get_sites_counts_slower}
\title{Count DNase or ATAC cuts around candidate sites}
\usage{
get_sites_counts_slower(
  sites,
  bam_file,
  shift_ATAC = FALSE,
  site_0base = TRUE,
  ncores = 10
)
}
\arguments{
\item{sites}{A data frame containing the candidate sites.}

\item{bam_file}{BAM file of DNase or ATAC-seq reads alignment}

\item{shift_ATAC}{Logical. When \code{shift_ATAC = TRUE},
it shifts reads aligned to the + strand by +4 bp,
and shifts reads aligned to the - strand by -5 bp.}

\item{site_0base}{Logical. If TRUE,
the coordinates of sites are 0-based as in BED format. Otherwise,
the coordinates of sites are 1-based.}

\item{ncores}{Number of cores available to parallel computing.}
}
\value{
A count matrix. The first half of the columns
are the read counts on the forward strand, and the second half of the
columns are the read counts on the reverse strand.
}
\description{
Count DNase or ATAC cuts around candidate sites on both strands.
Then combine the counts into one matrix, with the first half of the columns
representing the read counts on the forward strand,
and the second half of the columns representing the read counts
on the reverse strand with respect to the motif match location and strand.
}
\examples{
# Count DNase or ATAC cuts around candidate sites
sites_counts.mat <- get_sites_counts_slower(sites,
                                            bam_file = "ATAC_example.bam",
                                            shift_ATAC = TRUE,
                                            site_0base = TRUE,
                                            ncores = 10)
}
