% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_model_utils.R
\name{load_TOP_samples}
\alias{load_TOP_samples}
\title{Load TOP posterior samples}
\usage{
load_TOP_samples(TOP_samples_file, thin = 1, n.samples = 1000)
}
\arguments{
\item{TOP_samples_file}{file names of the posterior samples from all partitions}

\item{thin}{thinning rate of extract the posterior samples,
must be a positive integer.}

\item{n.samples}{Randomly choose n.samples posterior samples,
if the number of posterior samples is greater than n.samples}
}
\description{
Load TOP posterior samples
}
