% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mem.R
\name{ls.objects}
\alias{ls.objects}
\title{Title}
\usage{
ls.objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
\item{pos}{TBD}

\item{pattern}{TBD}

\item{order.by}{TBD}

\item{decreasing}{TBD}

\item{head}{TBD}

\item{n}{TBD}
}
\value{
TBD
}
\description{
Title
}
\keyword{internal}
