% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_utils.R
\name{match_tagcount_sites}
\alias{match_tagcount_sites}
\title{Match DNase tagcount matrices for candidate sites.}
\usage{
match_tagcount_sites(
  file_sites,
  file_dnase_tagcount_fwd,
  file_dnase_tagcount_rev,
  file_dnase_matrix_fwd,
  file_dnase_matrix_rev,
  path_bwtool = "bwtool"
)
}
\arguments{
\item{file_sites}{filename for candidate sites}

\item{file_dnase_tagcount_fwd}{filename for DNase tagcount in forward strand (Bigwig format)}

\item{file_dnase_tagcount_rev}{filename for DNase tagcount in reverse strand (Bigwig format)}

\item{file_dnase_matrix_fwd}{filename for DNase tagcount matrix in forward strand}

\item{file_dnase_matrix_rev}{filename for DNase tagcount matrix in reverse strand}

\item{path_bwtool}{Path to bwtool executable.}
}
\description{
Match DNase tagcount matrices for candidate sites.
}
