% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{merge_normalize_counts}
\alias{merge_normalize_counts}
\title{Merge counts from multiple replicates, then normalize merged counts}
\usage{
merge_normalize_counts(counts_files, idxstats_files, ref.size = 1e+08)
}
\arguments{
\item{counts_files}{DNase or ATAC counts files.}

\item{idxstats_files}{The idxstats files generated by samtools.}

\item{ref.size}{Scale to reference library size (Default: 1e8).}
}
\description{
Merge counts from multiple replicates, then normalize merged counts
}
