% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{merge_normalize_dnase_counts}
\alias{merge_normalize_dnase_counts}
\title{Merge DNase counts, then normalize merged DNase counts data.}
\usage{
merge_normalize_dnase_counts(
  dnase_counts_files,
  dnase_idxstats_files,
  ref.size = 1e+08
)
}
\arguments{
\item{dnase_counts_files}{DNase counts files}

\item{dnase_idxstats_files}{The DNase idxstats files generated by samtools}

\item{ref.size}{Scale to DNase-seq reference library size (Default: 100 million for DNase-seq)}
}
\description{
Merge DNase counts, then normalize DNase counts by library size scaling.
}
