% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPfunction.R
\name{MetaFixedPfunction_temp}
\alias{MetaFixedPfunction_temp}
\title{Title}
\usage{
MetaFixedPfunction_temp(icog_onco_score_infor_one, second.num)
}
\arguments{
\item{second.num}{}
}
\description{
Title
}
