% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPfunction.R
\name{MetaMixedPfunction}
\alias{MetaMixedPfunction}
\title{Title}
\usage{
MetaMixedPfunction(icog_onco_score_infor_one,
  icog_onco_score_infor_casecase_one, fixed.second.num, random.second.num)
}
\arguments{
\item{random.second.num}{}
}
\description{
Title
}
