% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVpolySelfDesign.R
\name{MvpolySelfDesign}
\alias{MvpolySelfDesign}
\title{Title}
\usage{
MvpolySelfDesign(
  y,
  x.self.design,
  z.design,
  baselineonly = NULL,
  additive = NULL,
  pairwise.interaction = NULL,
  saturated = NULL,
  missingTumorIndicator = 888,
  z.all = NULL,
  delta0 = NULL,
  cutoff = 10
)
}
\arguments{
\item{y}{}

\item{x.self.design}{}

\item{z.design}{}

\item{baselineonly}{}

\item{additive}{}

\item{pairwise.interaction}{}

\item{saturated}{}

\item{missingTumorIndicator}{}

\item{z.all}{}

\item{delta0}{}

\item{cutoff}{}
}
\description{
Title
}
