% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_utils.R
\name{normalize_bin_dnase}
\alias{normalize_bin_dnase}
\title{Normalize, binning and tranform DNase counts}
\usage{
normalize_bin_dnase(
  dnase_tagcounts,
  file_idxstats,
  type_bin = "M5",
  library_reference = 1e+08,
  transform = "log2"
)
}
\arguments{
\item{dnase_tagcounts}{DNase tagcount matrix}

\item{file_idxstats}{idxstats file generated by samtools}

\item{type_bin}{Type of MILLIPEDE binning scheme (Default: 'M5').}

\item{library_reference}{DNase-seq reference library size (Default: 100 million)}

\item{transform}{log2 or asinh transform}
}
\value{

}
\description{
Normalize DNase tagcounts by library size,
binning DNase using MILLIPEDE binning and then take log2 or asinh transform
}
