% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{normalize_bin_transform_counts}
\alias{normalize_bin_transform_counts}
\title{Normalize, bin and transform counts}
\usage{
normalize_bin_transform_counts(
  count_matrix,
  idxstats_file,
  ref.size = 1e+08,
  bin.method = c("M5", "M1", "M2", "M3", "M12", "M24"),
  transform = c("asinh", "log2")
)
}
\arguments{
\item{count_matrix}{A matrix of DNase (or ATAC) read counts.}

\item{idxstats_file}{The idxstats file generated by samtools}

\item{ref.size}{Scale to reference library size (Default: 1e8).}

\item{bin.method}{MILLIPEDE binning scheme (Default: 'M5').}

\item{transform}{Type of transformation for DNase (or ATAC) counts.
Options: asinh and log2.}
}
\value{
A data frame of normalized, binned and transformed counts.
}
\description{
Normalize counts by library size,
bin using MILLIPEDE binning method and then take asinh or log2 transform
}
