% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize DNase or ATAC counts}
\usage{
normalize_counts(counts, idxstats_file, ref.size = 1e+08)
}
\arguments{
\item{counts}{count matrix}

\item{idxstats_file}{The idxstats file generated by samtools}

\item{ref.size}{Scale to reference library size}
}
\description{
Normalize DNase or ATAC counts by library size scaling
}
