% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{normalize_dnase}
\alias{normalize_dnase}
\title{Normalize DNase counts}
\usage{
normalize_dnase(dnase_counts, idxstats_file, ref.size = 1e+08)
}
\arguments{
\item{dnase_counts}{DNase count matrix}

\item{idxstats_file}{The idxstats file generated by samtools}

\item{ref.size}{Scale to DNase-seq reference library size (Default: 100 million)}
}
\description{
Normalize DNase counts by library size scaling,
bin DNase using MILLIPEDE binning and then take log2 or asinh transform
}
