% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStepMleScoreSupport.R
\name{OneStepMleScoreTestSupport}
\alias{OneStepMleScoreTestSupport}
\title{Title}
\usage{
OneStepMleScoreTestSupport(y, baselineonly, additive, pairwise.interaction,
  saturated, missingTumorIndicator)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
