% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance.R
\name{performance_cpop2}
\alias{performance_cpop2}
\title{FUNCTION_TITLE}
\usage{
performance_cpop2(cpop_result, newx, covariates = NULL, s = "lambda.min")
}
\arguments{
\item{cpop_result}{PARAM_DESCRIPTION}

\item{newx}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: NULL}

\item{s}{PARAM_DESCRIPTION, Default: 'lambda.min'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
 # TODO
}
