% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot DNase or ATAC profiles}
\usage{
plot_profile(cuts, mlen = ncol(cuts)/2 - 200, title = "")
}
\arguments{
\item{cuts}{DNase or ATAC cuts}

\item{mlen}{motif length}

\item{title}{Title of the plot}
}
\description{
Plot DNase or ATAC profiles
}
