% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_profile_strands}
\alias{plot_profile_strands}
\title{Plot DNase or ATAC profiles separately by strands}
\usage{
plot_profile_strands(cuts, sites, mlen = 10, title = "")
}
\arguments{
\item{cuts}{DNase or ATAC cuts}

\item{sites}{candidate sites}

\item{mlen}{motif length}

\item{title}{Title of the plot}
}
\description{
Plot DNase or ATAC profiles separately by strands
}
