% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_norm}
\alias{predict_norm}
\title{Predict TF occupancy using posterior samples of regression coefficients}
\usage{
predict_norm(
  data,
  alpha_samples,
  beta_samples,
  tau_samples,
  sample = TRUE,
  average_parameters = FALSE,
  transform = "asinh"
)
}
\arguments{
\item{data}{A data frame or matrix. Columns are motif score and DNase features.
Rows are candidate sites.}

\item{alpha_samples}{posterior samples of alpha}

\item{beta_samples}{posterior samples of beta}

\item{tau_samples}{posterior samples of tau}

\item{sample}{logicals. If TRUE, samples from posterior predictions and then
take the mean of posterior prediction samples.}

\item{average_parameters}{logicals. If TRUE, uses the posterior mean of
regression coefficients to make predictions.}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP model. Default: transform = "asinh".}
}
\value{
The function returns a vector of predicted TF occupancy.
}
\description{
Predict TF occupancy using posterior samples of regression
coefficients trained from the TOP model
}
\examples{
predicted <- predict_norm(data, alpha_samples, beta_samples, tau_samples,
sample = TRUE, average_parameters = FALSE, transform = 'asinh')

}
