% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP_coef_mean}
\alias{predict_TOP_coef_mean}
\title{Predict TF occupancy using posterior mean of regression coefficients}
\usage{
predict_TOP_coef_mean(data, coef_mean, transform = "asinh")
}
\arguments{
\item{data}{A data frame or matrix. Columns are motif score and DNase features.
Rows are candidate sites.}

\item{coef_mean}{A numeric vector. The posterior mean of trained regression
coefficients, including the intercept and coefficients for motif score and
DNase features.
length(coef_mean) should be equal to 1+ncol(data).}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP model. Default: transform = "asinh".}
}
\value{
The function returns a vector of predicted TF occupancy.
}
\description{
Predict TF occupancy using posterior mean of regression
coefficients trained from the TOP model
}
