% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP_mean_coef}
\alias{predict_TOP_mean_coef}
\title{Predict TF occupancy using posterior mean of regression coefficients}
\usage{
predict_TOP_mean_coef(
  data,
  mean_coef,
  transform = c("asinh", "log2", "log", "none")
)
}
\arguments{
\item{data}{A data frame. Columns are motif score and DNase (or ATAC) bins.
Rows are candidate sites.}

\item{mean_coef}{A numeric vector. The posterior mean of trained regression
coefficients, including the intercept and coefficients for motif score and
DNase features.
length(mean_coef) should be equal to 1+ncol(data).}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP model. Options: asinh, log2, log, none.}
}
\value{
The function returns a vector of predicted TF occupancy.
}
\description{
Predict TF occupancy using posterior mean of regression
coefficients trained from the TOP model
}
