% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP_samples}
\alias{predict_TOP_samples}
\title{Predict TF occupancy using posterior samples of regression coefficients}
\usage{
predict_TOP_samples(
  data,
  TOP_samples,
  use_posterior_mean = FALSE,
  sample_predictions = TRUE,
  transform = c("asinh", "log2", "log", "none")
)
}
\arguments{
\item{data}{A data frame. Columns are motif score and DNase (or ATAC) bins.
Rows are candidate sites.}

\item{TOP_samples}{TOP posterior samples}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP model. Options: asinh, log2, log, none.}

\item{sample}{If TRUE, sample from posterior predictions and then
take the mean of posterior prediction samples.}

\item{average_parameters}{If TRUE, uses the posterior mean of
regression coefficients to make predictions.}
}
\value{
The function returns a vector of predicted TF occupancy.
}
\description{
Predict TF occupancy using posterior samples of regression
coefficients trained from the TOP model
}
