% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess_Frank.R
\name{PreProcess_Frank}
\alias{PreProcess_Frank}
\title{PreProcess_Frank}
\usage{
PreProcess_Frank(
  x_list,
  y_list,
  contrast = NULL,
  nFeatures = 50,
  combinationMethod = "OSP"
)
}
\arguments{
\item{x_list}{A list of data frames, with columns corresponding to features and rows corresponding to observations.}

\item{y_list}{A list of factor labels.}

\item{contrast}{A character vector describing which order of levels to contrast in y_list ("disease - control"), Default: NULL}

\item{nFeatures}{Number of features to return, Default: 50}

\item{combinationMethod}{Which p-value combination method to use, Default: 'OSP'}
}
\value{
A vector of feature names.
}
\description{
A function that implements feature selection from a list of data frames with corresponding labels.
}
\details{
contrast must be a character vector of length 1. If contrast is NULL, the first level of the first factor in y_list will be used as the reference level.
}
\examples{
 #EXAMPLE1
}
