% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbFitting.R
\name{ProbFitting}
\alias{ProbFitting}
\title{Title}
\usage{
ProbFitting(delta0, y, x.all, z.standard, z.all, missingTumorIndicator = NULL)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
