% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{process_candidate_sites}
\alias{process_candidate_sites}
\title{Full process to get candidate sites from FIMO result}
\usage{
process_candidate_sites(
  fimo_file = NULL,
  flank = 100,
  thresh_pValue = 1e-05,
  thresh_pwmscore = 0,
  blacklist_file = NULL,
  mapability_file = NULL,
  thresh_mapability = 0.8,
  out_file = NULL,
  bigWigAverageOverBed_path = "bigWigAverageOverBed"
)
}
\arguments{
\item{fimo_file}{Filename of FIMO result.}

\item{flank}{Flanking region (bp) around motif matches (default: 100)}

\item{thresh_pValue}{FIMO p-value threshold.}

\item{thresh_pwmscore}{FIMO PWM score threshold.}

\item{blacklist_file}{Filename of the blacklist regions}

\item{mapability_file}{Filename of the mapability reference file in bigWig format.}

\item{out_file}{Filename of processed candidate sites.}

\item{bigWigAverageOverBed_path}{Path to bigWigAverageOverBed executable.
This is only needed for computing mapability.}
}
\description{
Full process to get candidate sites from FIMO result
}
