% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{read_bam_cuts}
\alias{read_bam_cuts}
\title{Read DNase or ATAC cuts or coverage of cuts from BAM file}
\usage{
read_bam_cuts(
  bam_file,
  shift_ATAC = FALSE,
  shift_ATAC_bases = c(4L, -5L),
  return_type = c("cuts", "coverage")
)
}
\arguments{
\item{bam_file}{BAM file of DNase or ATAC-seq reads alignment}

\item{shift_ATAC}{Logical. When \code{shift_ATAC = TRUE},
it shifts reads according to \code{shift_ATAC_bases}.}

\item{shift_ATAC_bases}{Number of bases to shift. Default: c(4,-5),
shifts reads aligned to the + strand by +4 bp,
and shifts reads aligned to the - strand by -5 bp.}

\item{return_type}{Options for the returned data:
\dQuote{cuts} or \dQuote{coverage}.}
}
\value{
A GRange object of cuts or coverage
}
\description{
Read DNase or ATAC cuts or coverage of cuts from BAM file
}
