% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{readstats_total}
\alias{readstats_total}
\title{Get the number of total mapped reads from the idxstats file generated by samtools}
\usage{
readstats_total(
  file_idxstats,
  select_chr = TRUE,
  chr_list = paste0("chr", c(1:22))
)
}
\arguments{
\item{file_idxstats}{idxstats file generated by samtools.}

\item{select_chr}{whether to select chromosomes or use all chromosomes in the idxstats file.}

\item{chr_list}{Chromosomes to be included.}
}
\value{

}
\description{
Get the number of total mapped reads from the idxstats file generated by samtools
}
