% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rev_tagcount_bwtool}
\alias{rev_tagcount_bwtool}
\title{Flip the tagcounts generated from bwtool for motifs on the reverse (minus) strand}
\usage{
rev_tagcount_bwtool(file_sites, file_dnase_matrix_fwd, file_dnase_matrix_rev)
}
\arguments{
\item{file_sites}{filename for candidate sites}

\item{file_dnase_matrix_fwd}{filename for DNase tagcount matrix in forward strand}

\item{file_dnase_matrix_rev}{filename for DNase tagcount matrix in reverse strand}
}
\description{
Flip the tagcounts generated from bwtool for motifs on the reverse (minus) strand
}
