% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatterplot}
\alias{scatterplot}
\title{Simple scatter plot}
\usage{
scatterplot(
  x,
  y,
  xlab = "",
  ylab = "",
  title = "",
  xlim = NULL,
  ylim = NULL,
  color = "black"
)
}
\arguments{
\item{x}{x values (measured)}

\item{y}{y values (predicted)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{title of the figure}

\item{xlim}{range of x-axis values}

\item{ylim}{range of y-axis values}

\item{color}{color of the dots}
}
\description{
Simple scatter plot
}
