% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatterplot_predictions}
\alias{scatterplot_predictions}
\title{Make a scatter plot of measured and predicted occupancy}
\usage{
scatterplot_predictions(
  x,
  y,
  xlab = "measured",
  ylab = "predicted",
  title = "",
  transform = c("asinh", "log2", "sqrt"),
  xlim = c(0, 10),
  ylim = c(0, 10)
)
}
\arguments{
\item{x}{x values (measured)}

\item{y}{y values (predicted)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{title of the figure}

\item{transform}{transformation of the occupancy (asinh, log2 or sqrt)}

\item{xlim}{range of x-axis values}

\item{ylim}{range of y-axis values}
}
\description{
Make a scatter plot of measured and predicted occupancy
}
