% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTest.R
\name{ScoreMixedGlobalTestForAssoc}
\alias{ScoreMixedGlobalTestForAssoc}
\title{Title}
\usage{
ScoreMixedGlobalTestForAssoc(p.value.score.heter, score.baseline,
  infor.baseline)
}
\arguments{
\item{infor.baseline}{}
}
\description{
Title
}
