% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTest.R
\name{ScoreMixedGlobalTestForHeter}
\alias{ScoreMixedGlobalTestForHeter}
\title{Title}
\usage{
ScoreMixedGlobalTestForHeter(score.casecase, infor.casecase)
}
\arguments{
\item{infor.casecase}{}
}
\description{
Title
}
