% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTest.R
\name{ScoreTest}
\alias{ScoreTest}
\title{ScoreTest}
\usage{
ScoreTest(y, x, second.stage.structure = "additive",
  score.test.support = NULL, missingTumorIndicator = NULL)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
ScoreTest
}
