% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTestMixedModel.r
\name{ScoreTestMixedModel}
\alias{ScoreTestMixedModel}
\title{Title}
\usage{
ScoreTestMixedModel(y, x, z.design = z.design, score.test.support = NULL,
  missingTumorIndicator = NULL)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
