% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTestSelfDesign.R
\name{ScoreTestSelfDesign}
\alias{ScoreTestSelfDesign}
\title{Title}
\usage{
ScoreTestSelfDesign(y, x, z.design = z.design, score.test.support = NULL,
  missingTumorIndicator = NULL)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
