% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTestSupport.r
\name{ScoreTestSupport}
\alias{ScoreTestSupport}
\title{Title}
\usage{
ScoreTestSupport(y, baselineonly = NULL, additive = NULL,
  pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = NULL)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
