% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTestSupportSelfDesign.R
\name{ScoreTestSupportSelfDesign}
\alias{ScoreTestSupportSelfDesign}
\title{Title}
\usage{
ScoreTestSupportSelfDesign(y, x.self.design, z.design, baselineonly = NULL,
  additive = NULL, pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
