% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{select_features}
\alias{select_features}
\title{Extract PWM and bin features}
\usage{
select_features(data, pwm.name = "pwm", bin.name = "bin")
}
\arguments{
\item{data}{Input data frame}

\item{pwm.name}{column name (prefix) of the PMW score}

\item{bin.name}{column name (prefix) of the DNase or ATAC bins}
}
\description{
Extract PWM and bin features
}
