% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{select_model_coef_level}
\alias{select_model_coef_level}
\title{Select regression coefficients by the TOP hierarchy level}
\usage{
select_model_coef_level(
  TOP_mean_coef,
  tf_name,
  cell_type,
  level = c("best", "bottom", "middle", "top")
)
}
\arguments{
\item{TOP_mean_coef}{Trained TOP posterior mean regression coefficients.}

\item{tf_name}{specifies TF name.}

\item{cell_type}{specifies the cell type.}

\item{level}{Specific the TOP model hierarchy level to use.
Options: 'best', 'bottom', 'middle', or 'top'.
Default: 'best' -- choose the best model (lowest level of the hierarchy available):
If the TF motif and cell type is available in the training data,
then use the bottom level (TF- and cell-type-specific model).
otherwise, if TF motif (but not cell type) is available in the training data,
choose the middle level (TF-specific model) of that TF motif;
otherwise, use the top level TF-generic model.}
}
\value{
A list containing the results, including selected hierarchy level,
TOP model, regression coefficients (posterior mean).
}
\description{
Select regression coefficients by the TOP hierarchy level
}
