% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_models.R
\name{select_model_coef_mean}
\alias{select_model_coef_mean}
\title{Select TOP model hierarchy level}
\usage{
select_model_coef_mean(
  pwm_id,
  cell_type,
  tf_pwm_training_list,
  model_coefficients,
  model_level = c("best", "bottom", "middle", "top")
)
}
\arguments{
\item{pwm_id}{Motif PWM ID}

\item{cell_type}{Cell type}

\item{tf_pwm_training_list}{List of TF motifs with available coefficients trained.}

\item{model_coefficients}{Trained TOP model regression coefficients}

\item{model_level}{Specific the TOP model hierarchy level:
'top', 'bottom', 'middle', or 'best'.
Default: 'best', the lowest hierarchy level of the model will be used.}
}
\description{
If the TF motif is available in the training data and
model_level='best', then use the TF-specific model of that TF motif;
otherwise, use the top level TF-generic model.
}
