% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfFirstStageTest.R
\name{SelfFirstStageTest}
\alias{SelfFirstStageTest}
\title{Title}
\usage{
SelfFirstStageTest(beta, sigma, M, first.stage.mat, covar.names)
}
\arguments{
\item{first.stage.mat}{}
}
\description{
Title
}
