% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{simplenetworkPlot}
\alias{simplenetworkPlot}
\title{simplenetworkPlot}
\usage{
simplenetworkPlot(TOP_model, nFeatures = 50, s = "lambda.min")
}
\arguments{
\item{TOP_model}{A Transferable Omics Prediction model. THe output from the TOP_model function.}

\item{nFeatures}{The number of features that will be plotted. Default: 20}

\item{s}{Lambda value for the lasso model. Default is "lambda.min"}
}
\description{
simplenetworkPlot
}
\examples{
 # TODO: add examples.
}
