% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction_browser_tracks.R
\name{sort_merge_overlap_bedgraph}
\alias{sort_merge_overlap_bedgraph}
\title{Sort and merge the overlapping sites,
remove the overlapping sites with lower occupancy}
\usage{
sort_merge_overlap_bedgraph(predicted_bedgraph.df)
}
\arguments{
\item{predicted_bedgraph.df}{predicted TF occupancy in bedgraph format}
}
\description{
Sort and merge the overlapping sites,
remove the overlapping sites with lower occupancy
}
