% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepScoreTestSupportMixedModel.R
\name{StepScoreTestSupportMixedModel}
\alias{StepScoreTestSupportMixedModel}
\title{Title}
\usage{
StepScoreTestSupportMixedModel(delta0, y, x.all, z.standard, z.all)
}
\arguments{
\item{delta0}{}

\item{y}{}

\item{x.all}{}

\item{z.standard}{}

\item{z.all}{}
}
\description{
Title
}
