% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{sum_cuts_strand}
\alias{sum_cuts_strand}
\title{Count DNase cuts with different MILLIPEDE binning settings.}
\usage{
sum_cuts_strand(DNaseData, combine_strands = "vertical")
}
\arguments{
\item{DNaseData}{DNase data matrix, rows are candidate sites,
columns are DNase cuts with 100bp flanks around motifs in forward and reverse strands}

\item{combine_strands}{Method to combine DNase cuts from M24 bins in both strands to M12 bins:
'vertical' (default) or 'motif'.}
}
\value{

}
\description{
Count DNase cuts with different MILLIPEDE binning settings.
}
