% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialSetup.R
\name{SummaryResult}
\alias{SummaryResult}
\title{Title}
\usage{
SummaryResult(model.result, baselineonly, additive, pairwise.interaction,
  saturated, z.standard, covar.names, delta, z.design.additive,
  z.design.pairwise.interaction, z.design.saturated, tumor.names, z.all,
  x.self.design = NULL, z.design = NULL)
}
\arguments{
\item{z.design}{}
}
\description{
Title
}
