% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_Frank.R
\name{Surv_Frank}
\alias{Surv_Frank}
\title{FUNCTION_TITLE}
\usage{
Surv_Frank(
  x_list,
  y_list,
  nFeatures = 50,
  dataset_weights = NULL,
  sample_weights = FALSE,
  nCores = 1
)
}
\arguments{
\item{x_list}{a list of data frames, each containing the data for a single batch or dataset. Columns should be features and rows should be observations.}

\item{y_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
