% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_Frank.R
\name{Surv_Frank_Pred}
\alias{Surv_Frank_Pred}
\title{Create a prediction function.}
\usage{
Surv_Frank_Pred(coxnet_model, newx)
}
\arguments{
\item{coxnet_model}{PARAM_DESCRIPTION}

\item{newx}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
 #EXAMPLE1
}
