% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TakeoutIntercept.R
\name{TakeoutIntercept}
\alias{TakeoutIntercept}
\title{Title}
\usage{
TakeoutIntercept(delta, covariance.delta, M, tumor.names, z.all, covar.names)
}
\arguments{
\item{covar.names}{}
}
\description{
Title
}
