% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_survival.R
\name{TOP_survival}
\alias{TOP_survival}
\title{TOP_survival}
\usage{
TOP_survival(
  x_list,
  y_list,
  nFeatures = 50,
  dataset_weights = NULL,
  sample_weights = FALSE,
  nCores = 1
)
}
\arguments{
\item{x_list}{A list of data frames, each containing the data for a single batch or dataset. Columns should be features and rows should be observations.}

\item{y_list}{A list of data frames, where the first columns in each data frame is the time and the second column is the event status. The length of this list should be the same as the length of x_list.}
}
\value{
A cox net model
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
