% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_survival.R
\name{TOP_survivalPrediction}
\alias{TOP_survivalPrediction}
\title{TOP_survivalPrediction}
\usage{
TOP_survivalPrediction(TOP_survival, newx)
}
\arguments{
\item{TOP_survival}{A TOP_survival model. See \code{\link{TOP_survival}}.}

\item{newx}{A new dataset to predict the survival time.}
}
\value{
A vector of predicted survival time.
}
\description{
A prediction function for TOP_survival
}
\examples{
 #EXAMPLE1
}
