% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_tracks.R
\name{track_def_line}
\alias{track_def_line}
\title{Set genome browser track parameters for predictions in bedGraph format}
\usage{
track_def_line(
  tf_name,
  pwm_id,
  cell_type,
  rep_name,
  type_model,
  viewMax = 100,
  mycolor = "0,0,0"
)
}
\arguments{
\item{tf_name}{TF name}

\item{pwm_id}{PWM ID}

\item{cell_type}{cell type}

\item{rep_name}{replicate name}

\item{type_model}{model type}

\item{viewMax}{upper limit to view TF occupancy in genome browser}

\item{mycolor}{color of the track}
}
\value{

}
\description{
Set genome browser track parameters for predictions in bedGraph format
}
