% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TOP_models.R
\name{train_TOP_model}
\alias{train_TOP_model}
\title{Train TOP model for each partition separately}
\usage{
train_TOP_model(
  model_file,
  training_data_dir,
  training_data_name,
  logistic.model = FALSE,
  out_dir,
  partitions = 1:10,
  n.iter = 10000,
  n.burnin = 5000,
  n.chains = 3,
  n.thin = 10,
  transform = c("asinh", "log2", "sqrt", "none")
)
}
\arguments{
\item{model_file}{TOP logistic model file.}

\item{training_data_dir}{Directory for saving training data}

\item{training_data_name}{Prefix for training data file names}

\item{logistic.model}{If TRUE, use logistic version of the model}

\item{out_dir}{Output directory for TOP model posterior samples}

\item{partitions}{select which partition(s) to run}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.}

\item{n.chains}{number of Markov chains.}

\item{n.thin}{thinning rate, must be a positive integer.}

\item{transform}{Transformation of ChIP counts (asinh, log2, sqrt, or none)}
}
\description{
Train TOP model for each partition separately
}
