% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZDesigntoZall.R
\name{ZDesigntoZall}
\alias{ZDesigntoZall}
\title{Title}
\usage{
ZDesigntoZall(baselineonly, additive, pairwise.interaction, saturated,
  z.design.baselineonly, z.design.additive, z.design.pairwise.interaction,
  z.design.saturated)
}
\arguments{
\item{z.design.saturated}{}
}
\description{
Title
}
