% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZSelfDesigntoZall.R
\name{ZSelfDesigntoZall}
\alias{ZSelfDesigntoZall}
\title{Title}
\usage{
ZSelfDesigntoZall(x.self.design, baselineonly, additive, pairwise.interaction,
  saturated, z.design, z.design.baselineonly, z.design.additive,
  z.design.pairwise.interaction, z.design.saturated)
}
\arguments{
\item{x}{}
}
\description{
Title
}
