% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{add_chip_signals_to_sites}
\alias{add_chip_signals_to_sites}
\title{Add ChIP-seq signals to the candidate sites}
\usage{
add_chip_signals_to_sites(
  sites.df,
  chip_signal_file = NULL,
  chip_signal_sampleID = NULL,
  chip_signal_dir = "./",
  bigWigAverageOverBed_path = "bigWigAverageOverBed"
)
}
\arguments{
\item{sites.df}{data frame containing the candidate sites}

\item{chip_signal_file}{ChIP-seq signal file (.bigWig format)}

\item{chip_signal_sampleID}{ENCODE sample ID of ChIP-seq signals (.bigWig format)}

\item{chip_signal_dir}{Directory of ChIP-seq signals}

\item{bigWigAverageOverBed_path}{Path to bigWigAverageOverBed executable
(default: bigWigAverageOverBed)}
}
\description{
Add ChIP-seq signals to the candidate sites
}
