% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_training_data.R
\name{assemble_partition_training_data}
\alias{assemble_partition_training_data}
\title{Assemble TOP training data for all TF x cell type combos in one partition}
\usage{
assemble_partition_training_data(
  tf_cell_table,
  logistic.model = FALSE,
  chip_colname = "chip",
  training_chrs = paste0("chr", seq(1, 21, 2)),
  n.partitions = 10,
  part,
  max.sites = 10000,
  seed = 1
)
}
\arguments{
\item{tf_cell_table}{a data frame with TF name, cell type,
and links to the training data files.}

\item{logistic.model}{If TRUE, use logistic version of the model}

\item{chip_colname}{The column name of ChIP data in the combined data.}

\item{training_chrs}{Chromosomes used for training the model.}

\item{n.partitions}{split the data into partitions (default = 10)
to run Gibbs sampling in parallel.}

\item{part}{which partition to use for training the model}

\item{max.sites}{Max number of candidate sites in each partition.}

\item{seed}{seed used when sampling sites.}
}
\value{
Returns a data frame of training data with all TFs and cell type combos.
}
\description{
Assemble TOP training data for all TF x cell type combos in one partition
}
