% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_training_data.R
\name{assemble_partition_training_data}
\alias{assemble_partition_training_data}
\title{Assemble TOP training data for all TF x cell type combos in one partition}
\usage{
assemble_partition_training_data(
  tf_cell_table,
  logistic.model = FALSE,
  chip_colname = "chip",
  training_chrs = paste0("chr", seq(1, 21, 2)),
  n.partitions = 10,
  part,
  max.sites = 50000/n.partitions,
  seed = 123
)
}
\arguments{
\item{tf_cell_table}{a data frame with at least three columns:
TF name, cell type, and file name containing the corresponding training data.}

\item{logistic.model}{Logical; if TRUE, use the logistic version of TOP model.}

\item{chip_colname}{The column name of ChIP data in the combined data (default: "chip").}

\item{training_chrs}{Chromosomes used for training the model (default: odd chromosomes)}

\item{n.partitions}{Total number of partitions to split the training data (default: 10).}

\item{part}{specifies which partition to assemble the training data for}

\item{max.sites}{Max number of candidate sites in each partition (default: 50000/n.partitions).}

\item{seed}{seed used when sampling sites (default: 123).}
}
\description{
Assemble TOP training data for all TF x cell type combos in one partition
}
