% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_training_data.R
\name{assemble_TOP_training_data}
\alias{assemble_TOP_training_data}
\title{Assemble TOP training data for all TF x cell type combos,
then split training data into 10 partitions}
\usage{
assemble_TOP_training_data(
  tf_cell_table_file,
  logistic.model = FALSE,
  chip_colname = "chip",
  training_chrs = paste0("chr", seq(1, 21, 2)),
  n.partitions = 10,
  n.cores = n.partitions,
  max.sites = 50000,
  seed = 123
)
}
\arguments{
\item{tf_cell_table_file}{a tab delimited file with at least three columns:
The first three columns should be TF names, cell types, and the training data files.}

\item{logistic.model}{Logical; if TRUE, use the logistic version of TOP model.}

\item{chip_colname}{The column name of ChIP data in the combined data (default: "chip").}

\item{training_chrs}{Chromosomes used for training the model (default: odd chromosomes)}

\item{n.partitions}{Total number of partitions to split the training data (default: 10).}

\item{n.cores}{Number of cores to use in parallel
(default: equal to the number of partitions).}

\item{max.sites}{Max number of candidate sites in each partition (default: 50000/n.partitions).}

\item{seed}{seed used when sampling sites (default: 123).}
}
\description{
Assemble TOP training data for all TF x cell type combos,
then split training data into 10 partitions
}
