% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_training_data.R
\name{assemble_TOP_training_data}
\alias{assemble_TOP_training_data}
\title{Assemble TOP training data for all TF x cell type combos,
then split training data into partitions}
\usage{
assemble_TOP_training_data(
  tf_cell_table,
  transform = c("asinh", "log2", "sqrt", "none"),
  training_chrs = paste0("chr", seq(1, 21, 2)),
  sites_limit = Inf,
  total_partitions = 10,
  n_part = 1
)
}
\arguments{
\item{tf_cell_table}{a data frame with TF name, PWM name, cell type, and link to the training data with PWM, DNase and ChIP data.}

\item{transform}{Transform of ChIP counts. Options are 'asinh' (default), 'log2', 'sqrt', 'none'.}

\item{training_chrs}{Chromosomes used for training the model.}

\item{sites_limit}{Max number of candidate sites in each partition.}

\item{total_partitions}{split the data into partitions (default = 10) to run Gibbs sampling in parallel.}

\item{n_part}{which partition to use for training the model}
}
\value{
Returns a data frame of training data with all TFs and cell type combos.
}
\description{
Assemble TOP training data for all TF x cell type combos,
then split training data into partitions
}
