% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_training_data.R
\name{assemble_training_data}
\alias{assemble_training_data}
\title{Assemble TOP training data for all TF x cell type combos,
then split training data into 10 partitions}
\usage{
assemble_training_data(
  tf_cell_table,
  logistic.model = FALSE,
  chip_colname = "chip",
  training_chrs = paste0("chr", seq(1, 21, 2)),
  n.partitions = 10,
  n.cores = n.partitions,
  max.sites = 50000,
  seed = 1
)
}
\arguments{
\item{tf_cell_table}{a data frame with the first three columns as:
TF name, cell type, and file name containing the corresponding training data.}

\item{logistic.model}{Logical; if TRUE, use the logistic version of TOP model.}

\item{chip_colname}{The column name of ChIP data in the combined data
(default: "chip").}

\item{training_chrs}{Chromosomes used for training the model
(default: odd chromosomes)}

\item{n.partitions}{Total number of partitions to split the training data (default: 10).}

\item{n.cores}{Number of cores to use in parallel
(default: equal to the number of partitions).}

\item{max.sites}{Max number of candidate sites in each partition
(default: 50000/n.partitions).}

\item{seed}{seed used when sampling sites.}
}
\value{
A list of data frames, each containing one partition of the
training data set with all TF x cell type combos.
}
\description{
Prepare the training data for fitting TOP models.
It split training data into 10 partitions and
assemble training data for all TF x cell type combos for each of the partitions
using the function \code{assemble_partition_training_data}.
}
