% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{bam_sort_index_stats}
\alias{bam_sort_index_stats}
\title{Sort and index the BAM file, and retrieve stats (idxstats)}
\usage{
bam_sort_index_stats(
  bam_file,
  outdir,
  sort = TRUE,
  stats = TRUE,
  samtools_path = "samtools"
)
}
\arguments{
\item{bam_file}{Input BAM file.}

\item{outdir}{Output directory.}

\item{sort}{Logical. If TRUE, sort (and index) the BAM file.}

\item{stats}{Logical. If TRUE, retrieve stats (idxstats) in the index file.}

\item{samtools_path}{Path to samtools executable.}
}
\description{
Sort and index the BAM file, and retrieve stats (idxstats)
}
