% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{bam_sort_index_stats}
\alias{bam_sort_index_stats}
\title{Sort and index the BAM file, and retrieve the idxstats.}
\usage{
bam_sort_index_stats(
  bam_file,
  outdir = dirname(bam_file),
  sorted_bam_file,
  sort = TRUE,
  index = TRUE,
  idxstats = TRUE
)
}
\arguments{
\item{bam_file}{Input BAM file.}

\item{outdir}{Output directory (default: save to the directory of the BAM file).}

\item{sorted_bam_file}{Output file name for sorted BAM file.}

\item{sort}{Logical. If TRUE, sort the BAM file.}

\item{index}{Logical. If TRUE, index the BAM file.}

\item{idxstats}{Logical. If TRUE, retrieve idxstats from index file.}
}
\description{
Sort and index the BAM file, and
retrieve the idxstats using \code{Rsamtools}.
}
