% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{bin_transform_counts}
\alias{bin_transform_counts}
\title{Bin and transform count matrix}
\usage{
bin_transform_counts(
  counts,
  bin.method = c("M5", "M1", "M2", "M3", "M12", "M24"),
  transform = c("asinh", "log2")
)
}
\arguments{
\item{counts}{A matrix of DNase (or ATAC) read counts.}

\item{bin.method}{MILLIPEDE binning scheme (Default: 'M5').}

\item{transform}{Type of transformation for DNase (or ATAC) counts.
Options: asinh and log2.}
}
\value{
A data frame of binned and transformed counts.
}
\description{
Binning DNase (or ATAC) count matrix
using MILLIPEDE binning and then take asinh (or log2) transform
}
