% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dnase_atac_data.R
\name{bin_transform_counts}
\alias{bin_transform_counts}
\title{Bin and transform count matrix}
\usage{
bin_transform_counts(
  counts,
  bin_method = c("M5", "M24", "M12", "M3", "M2", "M1"),
  transform = c("asinh", "log2", "sqrt", "none")
)
}
\arguments{
\item{counts}{DNase or ATAC-seq read counts matrix}

\item{bin_method}{MILLIPEDE binning scheme (Default: \dQuote{M5}).}

\item{transform}{Type of transformation for DNase or ATAC counts.
Options: \dQuote{asinh}, \dQuote{log2}, \dQuote{sqrt}, \dQuote{none}.}
}
\value{
A data frame of binned and transformed counts.
}
\description{
Binning DNase or ATAC count matrix
using MILLIPEDE binning and then take asinh or log2 transform
}
