% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{calculateCCA}
\alias{calculateCCA}
\title{calculateCCA}
\usage{
calculateCCA(x_list)
}
\arguments{
\item{x_list}{a list of data.frames with identical features as columns.}
}
\value{
A square matrix of canonical correlation coefficients.
}
\description{
Returns a matrix of canonical correlation coefficients for a list of data.frames.
}
\examples{
 \dontrun{
 data(cpop_data_binary, package = "CPOP")

 x1 = cpop_data_binary$x1
 x2 = cpop_data_binary$x2
 x3 = cpop_data_binary$x3

 x_list <- list(x1,x2,x3)
 calculateCCA(x_list)
 }
}
