% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{coefNetworkPlot}
\alias{coefNetworkPlot}
\title{coefNetworkPlot}
\usage{
coefNetworkPlot(TOP_model, nFeatures = 20, s = "lambda.min")
}
\arguments{
\item{TOP_model}{A Transferable Omics Prediction model. THe output from the TOP_model function.}

\item{nFeatures}{The number of features that will be plotted. Default: 20}

\item{s}{Lambda value for the lasso model. Default is "lambda.min"}
}
\description{
coefNetworkPlot
}
\examples{
 # TODO: add examples.
}
