% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{compute_mapability}
\alias{compute_mapability}
\title{Get candidate sites by expanding FIMO motifs with flanking regions}
\usage{
compute_mapability(
  sites,
  file_mapability,
  path_bigWigAverageOverBed = "bigWigAverageOverBed"
)
}
\arguments{
\item{sites}{data frame of the candidate sites (BED format).}

\item{file_mapability}{Filename of the mapability reference file in bigWig format.}

\item{path_bigWigAverageOverBed}{Path to bigWigAverageOverBed executable.}
}
\value{

}
\description{
Get candidate sites by expanding FIMO motifs with flanking regions
}
