% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_genome_cuts}
\alias{count_genome_cuts}
\title{Count DNase-seq or ATAC-seq cuts along the genome}
\usage{
count_genome_cuts(
  bam_file,
  chrom_size_file,
  shift_ATAC = FALSE,
  outdir,
  outname,
  bedtools_path = "bedtools",
  bedGraphToBigWig_path = "bedGraphToBigWig"
)
}
\arguments{
\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File of genome sizes by chromosomes.}

\item{shift_ATAC}{Logical. When \code{shift_ATAC = TRUE},
it shifts reads aligned to the + strand by +4 bp,
and shifts reads aligned to the - strand by -5 bp.}

\item{outdir}{Output directory (default: use the directory of \code{bam_file}).}

\item{outname}{Output prefix (default: use the prefix of \code{bam_file}).}

\item{bedtools_path}{Path to \code{bedtools} executable.}

\item{bedGraphToBigWig_path}{Path to UCSC \code{bedGraphToBigWig} executable.}
}
\description{
Count genomic cleavage (5' end) from DNase-seq or
ATAC-seq BAM files using the \code{genomecov} tool from \code{bedtools}.
Save counts result in BigWig format using
the \code{bedGraphToBigWig} tool from UCSC.
For ATAC-seq, when \code{shift_ATAC = TRUE}, it shifts reads aligned to
the + strand by +4 bp, and shifts reads aligned to the - strand
by -5 bp (Buenrostro et al. 2013).
}
\examples{

# Obtain DNase-seq counts along the genome
count_genome_cuts(bam_file='DNase.bam',
                  chrom_size_file='hg38.chrom.sizes',
                  outdir='processed_data',
                  outname='K562.DNase',
                  bedtools_path='bedtools',
                  bedGraphToBigWig_path='bedGraphToBigWig')

# Obtain ATAC-seq counts along the genome
# We recommend using ATAC-seq reads that are pre-shifted and filtered.
# By default, shift_ATAC = FALSE, assuming ATAC-seq reads have been shifted.
count_genome_cuts(bam_file='ATAC.shifted.bam',
                  chrom_size_file='hg38.chrom.sizes',
                  shift_ATAC = FALSE,
                  outdir='processed_data',
                  outname='K562.ATAC',
                  bedtools_path='bedtools',
                  bedGraphToBigWig_path='bedGraphToBigWig')

# When shift_ATAC = TRUE, it shifts reads aligned to the + strand by +4 bp,
# and shifts reads aligned to the - strand by -5 bp.
count_genome_cuts(bam_file='ATAC.bam',
                  chrom_size_file='hg38.chrom.sizes',
                  shift_ATAC=TRUE,
                  outdir='processed_data',
                  outname='K562.ATAC',
                  bedtools_path='bedtools',
                  bedGraphToBigWig_path='bedGraphToBigWig')

# See more examples in \code{get_sites_counts()}.

}
