% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_genome_cuts}
\alias{count_genome_cuts}
\title{Count DNase-seq or ATAC-seq cuts along the genome}
\usage{
count_genome_cuts(
  bam_file,
  chrom_size_file,
  shift_ATAC = FALSE,
  shift_ATAC_bases = c(4L, -4L),
  outdir = dirname(bam_file),
  outname,
  bedtools_path = "bedtools",
  bedGraphToBigWig_path = "bedGraphToBigWig"
)
}
\arguments{
\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File of genome sizes by chromosomes.}

\item{shift_ATAC}{Logical. When \code{shift_ATAC = TRUE},
it shifts reads according to \code{shift_ATAC_bases}.}

\item{shift_ATAC_bases}{Number of bases to shift on + and - strands.}

\item{outdir}{Output directory (default: use the directory of \code{bam_file}).}

\item{outname}{Output prefix (default: use the prefix of \code{bam_file}).}

\item{bedtools_path}{Path to \code{bedtools} executable.}

\item{bedGraphToBigWig_path}{Path to UCSC \code{bedGraphToBigWig} executable.}
}
\description{
Count genomic cuts (5' end) from DNase-seq or
ATAC-seq BAM alignment files using \code{bedtools}
For ATAC-seq, when \code{shift_ATAC = TRUE}, reads
were shifted so as to address offsets and align the signal across strands.
}
\examples{
count_genome_cuts(bam_file='K562.ATAC.bam',
                  chrom_size_file='hg38.chrom.sizes',
                  shift_ATAC=TRUE,
                  outdir='processed_data',
                  outname='K562.ATAC.bam',
                  bedtools_path='bedtools',
                  bedGraphToBigWig_path='bedGraphToBigWig')
}
