% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{count_genome_cuts}
\alias{count_genome_cuts}
\title{Count DNase-seq or ATAC-seq cuts along the genome}
\usage{
count_genome_cuts(
  bam_file,
  chrom_size_file,
  shift_ATAC = FALSE,
  outdir = dirname(bam_file),
  outname,
  bedGraphToBigWig_path = "bedGraphToBigWig"
)
}
\arguments{
\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File of genome sizes by chromosomes.}

\item{shift_ATAC}{Logical. When \code{shift_ATAC = TRUE},
it shifts reads aligned to the + strand by +4 bp,
and shifts reads aligned to the - strand by -5 bp.}

\item{outdir}{Output directory (default: use the directory of \code{bam_file}).}

\item{outname}{Output prefix (default: use the prefix of \code{bam_file}).}

\item{bedGraphToBigWig_path}{Path to UCSC \code{bedGraphToBigWig} executable.#'}
}
\description{
Count genomic cleavage (5' end) from DNase-seq or
ATAC-seq BAM alignment files.
For ATAC-seq, when \code{shift_ATAC = TRUE}, it shifts reads aligned to
the + strand by +4 bp, and shifts reads aligned to the - strand
by -5 bp (Buenrostro et al. 2013).
This function treats paired end reads as independent.
If you have paired-end ATAC-seq reads and want to select the fragments
within NFR regions, please refer to the
\code{count_ATAC_genome_cuts_readpairs()} function.
}
