% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{count_normalize_chip}
\alias{count_normalize_chip}
\title{Count and normalize ChIP-seq read coverage}
\usage{
count_normalize_chip(
  sites.df,
  chip_bam_files,
  chip_idxstats_files,
  chrom_size_file,
  ref.size = 1e+07,
  transform = c("none", "asinh", "log2", "sqrt"),
  bedtools_path = "bedtools"
)
}
\arguments{
\item{sites.df}{data frame containing the candidate sites.}

\item{chip_bam_files}{ChIP-seq bam files.}

\item{chip_idxstats_files}{ChIP-seq idxstats files.
By default, use the corresponding '.idxstats.txt' files in the same directory
of the bam files.}

\item{chrom_size_file}{Chromosome size file.}

\item{ref.size}{ChIP-Seq reference library size (Default: 10 million)}

\item{transform}{Type of transformation for the ChIP read counts.
Options are 'none' (do not transform), 'asinh', 'log2', and 'sqrt'.}

\item{bedtools_path}{Path to bedtools executable (default: bedtools)}
}
\description{
Count and normalize ChIP-seq read coverage
}
