% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{CPOP_simplenetworkPlot}
\alias{CPOP_simplenetworkPlot}
\title{Title}
\usage{
CPOP_simplenetworkPlot(CPOP_model, nFeatures = 50, s = "lambda.min")
}
\arguments{
\item{CPOP_model}{A CPOP model}

\item{nFeatures}{The number of features that will be plotted}

\item{s}{lambda value. Default is "lambda.min"}
}
\description{
Title
}
\examples{
 # TODO: add examples.
}
