% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_95CI_p.R
\name{DisplayMixedScoreTestResult}
\alias{DisplayMixedScoreTestResult}
\title{Title}
\usage{
DisplayMixedScoreTestResult(score.baseline, infor.baseline, score.casecase,
  infor.casecase)
}
\arguments{
\item{infor.casecase}{}
}
\description{
Title
}
