% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_95CI_p.R
\name{DisplaySecondStageTestResult}
\alias{DisplaySecondStageTestResult}
\title{Title}
\usage{
DisplaySecondStageTestResult(logodds, sigma, self.design = F, places = 5)
}
\arguments{
\item{places}{}
}
\description{
Title
}
