% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmvpoly.R
\name{EMmvpoly}
\alias{EMmvpoly}
\title{Title}
\usage{
EMmvpoly(y, baselineonly = NULL, additive = NULL,
  pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888, delta0 = NULL)
}
\arguments{
\item{delta0}{}
}
\description{
Title
}
