% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmvpolySelfDesign.R
\name{EMmvpolySelfDesign}
\alias{EMmvpolySelfDesign}
\title{Title}
\usage{
EMmvpolySelfDesign(y, x.self.design, z.design, baselineonly = NULL,
  additive = NULL, pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888, z.all = NULL, delta0 = NULL)
}
\arguments{
\item{delta0}{}
}
\description{
Title
}
