% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMStepScoreTestSupportMixedModel.R
\name{EMStepScoreTestSupportMixedModel}
\alias{EMStepScoreTestSupportMixedModel}
\title{Title}
\usage{
EMStepScoreTestSupportMixedModel(delta0, y, x.all, z.standard, z.all,
  missingTumorIndicator)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
