% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_model_utils.R
\name{extract_TOP_mean_coef}
\alias{extract_TOP_mean_coef}
\title{Compute TOP posterior mean coefficients for all three levels of TOP model}
\usage{
extract_TOP_mean_coef(TOP_samples, tf_cell_combos, n.bins = 5)
}
\arguments{
\item{TOP_samples}{TOP samples combined from all partitions}

\item{tf_cell_combos}{a table of TF x cell type combinations}

\item{n.bins}{Number of DNase or ATAC bins in TOP model (default = 5)}
}
\value{

}
\description{
Compute TOP posterior mean coefficients for all three levels of TOP model
}
