% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filter_blacklist}
\alias{filter_blacklist}
\title{Filter out sites in blacklist regions}
\usage{
filter_blacklist(sites, file_blacklist, path_bedtools = "bedtools")
}
\arguments{
\item{sites}{data frame of the candidate sites (BED format).}

\item{file_blacklist}{Filename of the blacklist regions}

\item{path_bedtools}{Path to bedtools executable.}
}
\value{

}
\description{
Filter out sites in blacklist regions
}
