% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{filter_mapability}
\alias{filter_mapability}
\title{Title}
\usage{
filter_mapability(
  sites.df,
  mapability_file,
  thresh_mapability = 0.8,
  bigWigAverageOverBed_path = "bigWigAverageOverBed"
)
}
\arguments{
\item{sites.df}{A data frame of candidate binding sites with the first
6 columns the same as in the BED format}

\item{mapability_file}{ENCODE mapability bigWig file}

\item{thresh_mapability}{Mpability filter threshold (default: 0.8)}

\item{bigWigAverageOverBed_path}{path of bigWigAverageOverBed executable}
}
\description{
Title
}
