% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{fimo_motif_matches}
\alias{fimo_motif_matches}
\title{Run FIMO to scan for motif matches}
\usage{
fimo_motif_matches(
  motif_file,
  sequence_file,
  outname = "fimo.txt",
  outdir = dirname(outname),
  thresh_pValue = 1e-04,
  bfile = "--uniform--",
  skip_matched_sequence = TRUE,
  max_strand = FALSE,
  max_stored_scores = 1e+05,
  options = "",
  verbosity = 1,
  fimo_path = "fimo"
)
}
\arguments{
\item{motif_file}{Motif file.}

\item{sequence_file}{FASTA sequence file.}

\item{outname}{Output file name.}

\item{outdir}{Output directory.}

\item{thresh_pValue}{FIMO option 'thresh' for p-value threshold.}

\item{bfile}{FIMO option 'bfile' for background file or model.}

\item{skip_matched_sequence}{FIMO option 'skip_matched_sequence'.
Turns off output of the sequence of motif matches.
This speeds up processing considerably.}

\item{max_strand}{FIMO option 'max_strand'. If matches on both strands at a
given position satisfy the output threshold,
only report the match for the strand with the higher score.
If the scores are tied, the matching strand is chosen at random.}

\item{max_stored_scores}{The maximum number of stored matches.}

\item{options}{Other options for FIMO.}

\item{verbosity}{A number that regulates the verbosity level of the output
information messages.
If set to 1 (quiet) then it will only output error messages whereas the
other extreme 5 (dump) outputs lots of mostly useless information.}

\item{fimo_path}{Path to FIMO command line executable.}
}
\description{
Run FIMO to scan for motif matches along the genome.
This is a wrapper function to run FIMO command line.
}
\examples{
fimo_motif_matches(motif_file='MA0139.1.meme',
                   sequence_file='hg38.fa',
                   thresh_pValue=1e-5,
                   outname='MA0139.1_1e-5.fimo.txt',
                   fimo_path='fimo')
}
