% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TOP_model.R
\name{fit_TOP_jags}
\alias{fit_TOP_jags}
\title{Fit TOP model using JAGS}
\usage{
fit_TOP_jags(
  combined_data,
  parameters.to.save,
  model.file,
  n.iter = 1000,
  n.burnin = 100,
  n.thin = 2,
  n.chains = 1,
  DIC = TRUE
)
}
\arguments{
\item{combined_data}{combined training data.}

\item{parameters.to.save}{character vector of the names of the parameters to
save which should be monitored.}

\item{model.file}{file containing the model written in BUGS code.}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.}

\item{n.thin}{thinning rate, must be a positive integer (default=2).}

\item{n.chains}{number of Markov chains (default: 1).}

\item{DIC}{logical; if TRUE (default), compute deviance, pD, and DIC.}
}
\value{

}
\description{
Fit TOP model using JAGS
}
