% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TOP_models.R
\name{fit_TOP_logistic_M5_model_jags}
\alias{fit_TOP_logistic_M5_model_jags}
\title{Fit TOP logistic model with M5 DNase (or ATAC) bins and binary ChIP labels.}
\usage{
fit_TOP_logistic_M5_model_jags(
  data,
  model.file,
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  n.chains = 3,
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  quiet = FALSE
)
}
\arguments{
\item{data}{a data frame containing the combined training data.}

\item{model.file}{TOP model file written in JAGS.}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.
Default is n.iter/2, that is, discarding the first half of the simulations.}

\item{n.chains}{number of Markov chains (default: 3).}

\item{n.thin}{thinning rate, must be a positive integer.
Default is max(1, floor(n.chains * (n.iter-n.burnin) / 1000))
which will only thin if there are at least 2000 simulations.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}
}
\value{
A jagsfit object from \code{R2jags}.
}
\description{
Fit TOP logistic model with M5 DNase (or ATAC) bins and
binary ChIP labels using JAGS.
The R2jags package \code{\link[R2jags]{R2jags}} is required.
}
