% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TOP_models.R
\name{fit_TOP_logistic_M5_model_jags}
\alias{fit_TOP_logistic_M5_model_jags}
\title{Fit TOP binary (logistic) model with M5 bins using JAGS}
\usage{
fit_TOP_logistic_M5_model_jags(
  data,
  model.file,
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  n.chains = 3,
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  quiet = FALSE
)
}
\arguments{
\item{data}{a data frame containing the combined training data.}

\item{model.file}{file containing the TOP model written in BUGS code.}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.
Default is n.iter/2, that is, discarding the first half of the simulations.}

\item{n.chains}{number of Markov chains (default: 3).}

\item{n.thin}{thinning rate, must be a positive integer.
Default is max(1, floor(n.chains * (n.iter-n.burnin) / 1000))
which will only thin if there are at least 2000 simulations.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}
}
\description{
Fit TOP binary (logistic) model with M5 bins using JAGS
}
