% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TOP_models.R
\name{fit_TOP_M5_model_jags}
\alias{fit_TOP_M5_model_jags}
\title{Fit TOP quantitative occupancy model with M5 bins}
\usage{
fit_TOP_M5_model_jags(
  data,
  model_file,
  transform = c("asinh", "log2", "sqrt", "none"),
  n_iter = 2000,
  n_burnin = floor(n_iter/2),
  n_chains = 3,
  n_thin = max(1, floor((n_iter - n_burnin)/1000)),
  quiet = FALSE
)
}
\arguments{
\item{data}{a data frame containing the combined training data.}

\item{model_file}{TOP model file written in JAGS.}

\item{transform}{Type of transformation for ChIP counts.
Possible values are "asinh", "log2", "sqrt", and "none" (no transformation).}

\item{n_iter}{number of total iterations per chain (including burn in).}

\item{n_burnin}{length of burn in, i.e. number of iterations to discard at the beginning.
Default is n_iter/2, that is, discarding the first half of the simulations.}

\item{n_chains}{number of Markov chains (default: 3).}

\item{n_thin}{thinning rate, must be a positive integer.
Default is max(1, floor(n_chains * (n_iter-n_burnin) / 1000))
which will only thin if there are at least 2000 simulations.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}
}
\value{
A jagsfit object from \code{R2jags}.
}
\description{
Fit TOP quantitative occupancy model with M5 bins using JAGS.
The R2jags package \code{\link[R2jags]{R2jags}} is required.
}
