% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TOP_models.R
\name{fit_TOP_model}
\alias{fit_TOP_model}
\title{Fit TOP model for each partition separately}
\usage{
fit_TOP_model(
  all_training_data,
  all_training_data_files,
  model.file,
  logistic.model = FALSE,
  out.dir = "TOP_samples",
  transform = c("asinh", "log2", "sqrt", "none"),
  partitions = 1:10,
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  n.chains = 3,
  n.thin = max(1, floor((n.iter - n.burnin)/1000))
)
}
\arguments{
\item{all_training_data}{Training data of all partitions.}

\item{all_training_data_files}{Training data files of all partitions
Requires either all_training_data or all_training_data_files.}

\item{model.file}{TOP logistic model file.}

\item{logistic.model}{If TRUE, use logistic version of the model}

\item{out.dir}{Output directory for TOP model posterior samples}

\item{transform}{Transformation of ChIP counts (asinh, log2, sqrt, or none).
Only needed when logistic.model is FALSE.}

\item{partitions}{select which partition(s) to run}

\item{n.iter}{number of total iterations per chain (including burn in).}

\item{n.burnin}{length of burn in, i.e. number of iterations to discard at the beginning.}

\item{n.chains}{number of Markov chains.}

\item{n.thin}{thinning rate, must be a positive integer.}
}
\description{
Fit TOP model for each partition separately
}
