% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flank_fimo_sites}
\alias{flank_fimo_sites}
\title{Get candidate sites by expanding FIMO motifs with flanking regions}
\usage{
flank_fimo_sites(file_fimo, flank = 100)
}
\arguments{
\item{file_fimo}{Filename of FIMO results.}

\item{flank}{Size (bp) of flanking region on each side of motif. Default: 100.}
}
\value{

}
\description{
Get candidate sites by expanding FIMO motifs with flanking regions
}
