% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{flank_fimo_sites}
\alias{flank_fimo_sites}
\title{Get candidate sites using FIMO motifs with flanking regions}
\usage{
flank_fimo_sites(fimo_file, flank = 100)
}
\arguments{
\item{fimo_file}{FIMO result .txt file}

\item{flank}{Flanking region (bp) around motif matches (default: 100)}
}
\description{
Get candidate sites using FIMO motifs with flanking regions
}
