% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{flip_rev_strand_counts}
\alias{flip_rev_strand_counts}
\title{Flip the counts generated from bwtool for motifs on the reverse (minus) strand}
\usage{
flip_rev_strand_counts(
  sites.df,
  fwd_matrix_file,
  rev_matrix_file,
  write = FALSE
)
}
\arguments{
\item{sites.df}{data frame containing candidate sites}

\item{fwd_matrix_file}{File for count matrix in forward strand}

\item{rev_matrix_file}{File for count matrix in reverse strand}

\item{write}{Logical; if TRUE, write count matrices to fwd_matrix_file and rev_matrix_file}
}
\description{
Flip the counts generated from bwtool for motifs on the reverse (minus) strand
}
