% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{get_sites_counts}
\alias{get_sites_counts}
\title{Get count matrices around candidate sites}
\usage{
get_sites_counts(
  sites.df,
  genomecount_dir,
  genomecount_name,
  tmpdir = genomecount_dir,
  bwtool_path = "bwtool"
)
}
\arguments{
\item{sites.df}{data frame containing candidate sites}

\item{genomecount_dir}{Directory for genome counts}

\item{genomecount_name}{File prefix for genome counts}

\item{tmpdir}{temporary directory to save intermediate files, default: genomecount_dir}

\item{bwtool_path}{Path to bwtool executable.}
}
\description{
Get count matrices around candidate sites
}
