% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{get_sites_counts}
\alias{get_sites_counts}
\title{Get count matrices around candidate binding sites}
\usage{
get_sites_counts(
  sites,
  bam_file,
  chrom_size_file,
  adjust_ATACshift = FALSE,
  genomecount_dir,
  genomecount_name,
  tmpdir = genomecount_dir,
  bedGraphToBigWig_path = "bedGraphToBigWig",
  bwtool_path = "bwtool"
)
}
\arguments{
\item{sites}{A data frame containing the candidate sites.}

\item{bam_file}{Sorted BAM file.}

\item{chrom_size_file}{File of genome sizes by chromosomes.}

\item{adjust_ATACshift}{Logical. If TRUE, adjust the candidate site windows (by 1bp)
for motif matches on the - strand due to the shift of ATAC-seq reads.}

\item{genomecount_dir}{Directory for genome counts.}

\item{genomecount_name}{File prefix for genome counts.}

\item{tmpdir}{Temporary directory to save intermediate files.}

\item{bedGraphToBigWig_path}{Path to UCSC \code{bedGraphToBigWig} executable.#'}

\item{bwtool_path}{Path to \code{bwtool} executable.}
}
\value{
A count matrix. The first half of the columns
are the read counts on the forward strand, and the second half of the
columns are the read counts on the reverse strand.
}
\description{
Extract counts around candidate binding sites on both strands
from the genome counts (BigWig files prepared using \code{count_genome_cuts}).
It utilizes the \code{extract bed} function from the \code{bwtool} software
to extract the read counts.
Then combine the counts into one matrix, with the first half of the columns
representing the read counts on the forward strand,
and the second half of the columns representing the read counts
on the reverse strand.
}
\examples{
# Get ATAC-seq count matrices around candidate sites
sites_counts.mat <- get_sites_counts(sites,
                                     bam_file='K562.ATAC.bam',
                                     chrom_size_file,
                                     adjust_ATACshift=TRUE,
                                     genomecount_dir='processed_data',
                                     genomecount_name='K562.ATAC',
                                     bedGraphToBigWig_path='bedGraphToBigWig',
                                     bwtool_path='bwtool')
}
