% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{get_total_reads}
\alias{get_total_reads}
\title{Get total number of mapped reads from the idxstats file generated by samtools}
\usage{
get_total_reads(
  idxstats_file,
  select.chr = TRUE,
  chrs = paste0("chr", c(1:22))
)
}
\arguments{
\item{idxstats_file}{idxstats file generated by samtools.}

\item{select.chr}{If TRUE, use all chromosomes in the idxstats file.
If FALSE, select chromosomes in chrs.}

\item{chrs}{Chromosomes to be included.}
}
\description{
Get total number of mapped reads from the idxstats file generated by samtools
}
