% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{get_total_reads}
\alias{get_total_reads}
\title{Get total number of mapped reads from the idxstats file generated by samtools}
\usage{
get_total_reads(
  idxstats_file,
  select.chr = TRUE,
  chrs = paste0("chr", c(1:22))
)
}
\arguments{
\item{idxstats_file}{idxstats file generated by samtools.}

\item{select.chr}{Logitic. If TRUE (default), select chromosomes in chrs.
Otherwise, use all chromosomes in the idxstats file.}

\item{chrs}{Chromosomes to be included (default: paste0('chr', c(1:22)))}
}
\description{
Get total number of mapped reads from the idxstats file generated by samtools
}
