% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{get_total_reads}
\alias{get_total_reads}
\title{Get total number of mapped reads from the idxstats file}
\usage{
get_total_reads(
  idxstats_file,
  select_chr = TRUE,
  chrs = paste0("chr", c(1:22))
)
}
\arguments{
\item{idxstats_file}{idxstats file.}

\item{select_chr}{Logitic. If TRUE, count reads in
chromosomes specified in \code{chrs}.
Otherwise, use all chromosomes in the idxstats file.}

\item{chrs}{Chromosomes to be included (default: chr1, ..., chr22).}
}
\value{
The total number of mapped reads.
}
\description{
Loads the idxstats file, and count
the total number of mapped reads generated by \code{bam_sort_index_stats}.
}
