% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialSetup.R
\name{InitialSetup}
\alias{InitialSetup}
\title{Title}
\usage{
InitialSetup(y.pheno.complete, baselineonly, additive, pairwise.interaction,
  saturated, x.self.design = NULL, z.design = NULL)
}
\arguments{
\item{z.design}{}
}
\description{
Title
}
\examples{


}
