% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{merge_normalize_bin_transform_counts}
\alias{merge_normalize_bin_transform_counts}
\title{Merge counts from multiple replicates, then normalize,
bin and transform the merged counts}
\usage{
merge_normalize_bin_transform_counts(
  counts_files,
  idxstats_files,
  ref.size = 1e+08,
  bin.method = c("M5", "M1", "M2", "M3", "M12", "M24"),
  transform = c("asinh", "log2")
)
}
\arguments{
\item{counts_files}{DNase or ATAC-seq read counts matrix files.}

\item{idxstats_files}{The idxstats files generated by samtools.}

\item{ref.size}{Scale to DNase-seq or ATAC-seq reference library size (Default: 1e8).}

\item{bin.method}{MILLIPEDE binning scheme (Default: 'M5').}

\item{transform}{Type of transformation for DNase (or ATAC) counts.
Options: asinh and log2.}
}
\value{
A data frame of merged, normalized, binned, and transformed counts.
}
\description{
Merge (sum) read counts from multiple replicate samples,
normalize the read counts by scaling to the reference library size,
then bin and transform the merged counts.
}
