% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{merge_normalize_counts}
\alias{merge_normalize_counts}
\title{Merge DNase or ATAC-seq counts from multiple replicates,
then normalize merged counts.}
\usage{
merge_normalize_counts(counts_files, idxstats_files, ref_size = 1e+08)
}
\arguments{
\item{counts_files}{DNase or ATAC-seq read counts matrix files.}

\item{idxstats_files}{The idxstats files generated by \code{samtools}.}

\item{ref_size}{Scale to reference library size (default: 1e8).}
}
\value{
A matrix of merged and normalized counts.
}
\description{
Merge DNase or ATAC-seq read counts from multiple replicate samples,
and normalize the read counts by scaling to the reference library size.
}
