% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{millipede_binning}
\alias{millipede_binning}
\title{MILLIPEDE binning}
\usage{
millipede_binning(counts, combine_strands = c("vertical", "motif"))
}
\arguments{
\item{counts}{Counts matrix, rows are candidate sites,
columns are DNase or ATAC counts with 100bp flanks around motifs in forward and reverse strands}

\item{combine_strands}{Method to combine counts from M24 bins in both strands to M12 bins:
'vertical' (default) or 'motif'.}
}
\description{
MILLIPEDE binning
}
