% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVpolySelfDesign.R
\name{MvpolySelfDesign}
\alias{MvpolySelfDesign}
\title{Title}
\usage{
MvpolySelfDesign(y, baselineonly, additive, pairwise.interaction, saturated,
  delta0 = NULL, cutoff = 10)
}
\arguments{
\item{y}{}

\item{baselineonly}{}

\item{additive}{}

\item{pairwise.interaction}{}

\item{saturated}{}

\item{delta0}{}

\item{cutoff}{}
}
\description{
Title
}
