% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{normalize_bin_dnase}
\alias{normalize_bin_dnase}
\title{Normalize, binning and transform DNase counts}
\usage{
normalize_bin_dnase(
  dnase_counts,
  idxstats_file,
  ref.size = 1e+08,
  bin.method = c("M5", "M1", "M2", "M3", "M12", "M24"),
  transform = c("asinh", "log2")
)
}
\arguments{
\item{dnase_counts}{DNase count matrix}

\item{idxstats_file}{The idxstats file generated by samtools}

\item{ref.size}{Scale to DNase-seq reference library size (Default: 100 million)}

\item{bin.method}{MILLIPEDE binning scheme (Default: 'M5').}

\item{transform}{asinh or log2 transform of DNase counts.}
}
\description{
Normalize DNase counts by library size,
binning DNase using MILLIPEDE binning and then take asinh or log2 transform
}
