% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{normalize_chip}
\alias{normalize_chip}
\title{Normalize (and transform) ChIP-seq read coverage}
\usage{
normalize_chip(
  chip_counts,
  idxstats_file,
  ref.size = 1e+07,
  transform = c("none", "asinh", "log2", "sqrt")
)
}
\arguments{
\item{chip_counts}{A vector of ChIP-seq counts.}

\item{idxstats_file}{The idxstats file generated by \code{samtools}.}

\item{ref.size}{ChIP-Seq reference library size (Default: 10 million).}

\item{transform}{Type of transformation for the ChIP read counts.
Options are 'none' (no transformation), 'asinh', 'log2', and 'sqrt'.}
}
\value{
A vector of transformed ChIP-seq counts.
}
\description{
Normalize the ChIP-seq read counts by scaling
to the reference ChIP-seq library size,
and transform the ChIP-seq read counts.
}
