% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_utils.R
\name{normalize_chip}
\alias{normalize_chip}
\title{Normalize and transform ChIP counts}
\usage{
normalize_chip(
  chip_counts,
  idxstats_file,
  ref.size = 1e+07,
  transform = "asinh"
)
}
\arguments{
\item{chip_counts}{ChIP-seq counts}

\item{idxstats_file}{idxstats file generated by samtools}

\item{ref.size}{ChIP-Seq reference library size (Default: 10 million)}

\item{transform}{log2 or asinh transform}
}
\description{
Normalize and transform ChIP counts
}
