% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{normalize_chip}
\alias{normalize_chip}
\title{Normalize, binning and tranform DNase counts}
\usage{
normalize_chip(
  chip_counts,
  file_idxstats,
  library_reference = 1e+07,
  transform = "asinh"
)
}
\arguments{
\item{chip_counts}{ChIP-seq counts}

\item{file_idxstats}{idxstats file generated by samtools}

\item{library_reference}{ChIP-Seq reference library size (Default: 10 million)}

\item{transform}{log2 or asinh transform}
}
\value{

}
\description{
Normalize, binning and tranform DNase counts
}
