% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_atac_utils.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize read counts}
\usage{
normalize_counts(counts, idxstats_file, ref_size = 1e+08)
}
\arguments{
\item{counts}{DNase or ATAC-seq read counts matrix}

\item{idxstats_file}{The \code{idxstats} file generated by samtools.}

\item{ref_size}{Normalize to reference library size (default: 1e8).}
}
\value{
A matrix of normalize read counts.
}
\description{
Normalize DNase or ATAC-seq read counts by library sizes.
It first obtain the total mapped reads from the current sample, and
then scales the read counts for the current data to a reference library size.
}
