% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_profile_strands}
\alias{plot_profile_strands}
\title{Plot DNase or ATAC profiles by strands of motif matches}
\usage{
plot_profile_strands(
  cuts,
  sites,
  mlen = ncol(cuts)/2 - 200,
  title = "",
  strand = c("both", "+", "-")
)
}
\arguments{
\item{cuts}{Matrix of DNase or ATAC cuts}

\item{sites}{Data frame of candidate sites}

\item{mlen}{Motif length}

\item{title}{Title of the plot}
}
\description{
Plot DNase or ATAC profiles by strands of motif matches
}
\examples{
\dontrun{
plot_profile_strands(cuts, sites, title='Footprint profiles around binding sites')
}
}
