% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frankenstein_CPOP.R
\name{predict_cpop2}
\alias{predict_cpop2}
\title{Prectict using the CPOP2 results.}
\usage{
predict_cpop2(cpop_result, newx, covariates = NULL, s = "lambda.min")
}
\arguments{
\item{cpop_result}{PARAM_DESCRIPTION}

\item{newx}{PARAM_DESCRIPTION}

\item{covariates}{PARAM_DESCRIPTION, Default: NULL}

\item{s}{PARAM_DESCRIPTION, Default: 'lambda.min'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
