% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_model.R
\name{predict_TOP}
\alias{predict_TOP}
\title{Prectict using the Trasferable Omics Prediction model.}
\usage{
predict_TOP(TOP_model, newx, covariates = NULL, s = "lambda.min")
}
\arguments{
\item{TOP_model}{The output from the TOP_model function.}

\item{newx}{A matrix of the new data to be predicted. The columns should be features and the rows should be samples.}

\item{covariates}{A data frame of the same covariates that were used in the CPOP2 model, Default: NULL}

\item{s}{Lambda value for the lasso model, Default: 'lambda.min'}
}
\value{
A vector of predictions for the new data.
}
\description{
A prediction function for the Trasferable Omics Prediction model.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
