% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP}
\alias{predict_TOP}
\title{Predict quantitative TF occupancy or TF binding probability}
\usage{
predict_TOP(
  data,
  tf_name,
  cell_type,
  TOP_coef,
  level = c("best", "bottom", "middle", "top"),
  logistic.model = FALSE,
  transform = c("asinh", "log2", "log", "none")
)
}
\arguments{
\item{data}{A data frame containing motif PWM score and DNase (or ATAC) bins.}

\item{tf_name}{specifies TF name.}

\item{cell_type}{specifies the cell type.}

\item{TOP_coef}{A list containing the posterior mean of TOP regression coefficients.}

\item{level}{Specific the TOP model hierarchy level to use.
Options: 'best', 'bottom', 'middle', or 'top'.
Default: 'best' -- choose the best model (lowest level of the hierarchy available):
If the TF motif and cell type is available in the training data,
then use the bottom level (TF- and cell-type-specific model).
otherwise, if TF motif (but not cell type) is available in the training data,
choose the middle level (TF-specific model) of that TF motif;
otherwise, use the top level TF-generic model.}

\item{logistic.model}{Logical; if TRUE, use the logistic version of TOP model.}

\item{transform}{Type of transformation for ChIP counts.
Possible values are "asinh", "log2", "sqrt", and "none" (no transformation).
Only needed when logistic.model is FALSE.}
}
\description{
Predict quantitative TF occupancy or TF binding probability
using TOP model trained from ChIP-seq read counts or binary labels.
}
