% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP}
\alias{predict_TOP}
\title{Predict quantitative TF occupancy or TF binding probability}
\usage{
predict_TOP(
  data,
  TOP_coef,
  logistic.model = FALSE,
  posterior.option = c("mean", "samples"),
  transform = c("asinh", "log2", "log", "none")
)
}
\arguments{
\item{data}{A data frame containing motif PWM score and DNase (or ATAC) bins.}

\item{TOP_coef}{A data frame or list containing the posterior samples
or posterior mean of TOP regression coefficients.}

\item{logistic.model}{Logical; if TRUE, use the logistic version of TOP model.}

\item{posterior.option}{Method to predict occupancy.
'samples': uses posterior samples,
'mean': uses posterior mean of regression coefficients.}

\item{transform}{Type of transformation for ChIP counts.
Possible values are "asinh", "log2", "sqrt", and "none" (no transformation).
Only needed when logistic.model is FALSE.}
}
\value{
Returns a vector of predicted TF occupancy (posterior mean).
}
\description{
Predict quantitative TF occupancy or TF binding probability
using TOP model trained from ChIP-seq read counts or binary labels.
}
