% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP}
\alias{predict_TOP}
\title{Predict quantitative TF occupancy or TF binding probability}
\usage{
predict_TOP(
  data,
  TOP_model,
  logistic.model = FALSE,
  posterior.option = c("samples", "mean"),
  transform = c("asinh", "log2", "log", "none")
)
}
\arguments{
\item{data}{A data frame. Columns are motif score and DNase (or ATAC) bins.
Rows are candidate sites.}

\item{TOP_model}{TOP posterior samples or posterior mean of regression
coefficients.}

\item{posterior.option}{"samples": uses posterior samples, "mean": uses posterior mean of
trained regression coefficients}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP quantitative model. Options: asinh, log2, log, none.
This does not apply to the logistic version.}
}
\value{
The function returns a vector of predicted TF occupancy.
}
\description{
Predict quantitative TF occupancy or TF binding probability
using TOP model trained from ChIP-seq read counts or binary labels.
}
