% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP_logistic_mean_coef}
\alias{predict_TOP_logistic_mean_coef}
\title{Predict TF binding probability by TOP logistic model
with posterior mean of regression coefficients}
\usage{
predict_TOP_logistic_mean_coef(data, mean_coef)
}
\arguments{
\item{data}{A data frame. Columns are motif score and DNase (or ATAC) bins.
Rows are candidate sites.}

\item{mean_coef}{A numeric vector. The posterior mean of trained regression
coefficients, including the intercept and coefficients for motif score and
DNase features.
length(mean_coef) should be equal to 1+ncol(data).}
}
\value{
The function returns a vector of predicted TF binding probabilities
}
\description{
Predict TF binding probability using posterior mean of the regression
coefficients trained from TOP logistic model
}
