% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP_mean_coef}
\alias{predict_TOP_mean_coef}
\title{Predict TF occupancy using posterior mean of regression coefficients}
\usage{
predict_TOP_mean_coef(
  data,
  mean_coef,
  transform = c("asinh", "log2", "log", "none")
)
}
\arguments{
\item{data}{A data frame containing motif PWM score and DNase (or ATAC) bins.}

\item{mean_coef}{A numeric vector. The posterior mean of trained regression
coefficients, including the intercept and coefficients for PWM score and
DNase (or ATAC) bins.
length(mean_coef) should be equal to 1+ncol(data).}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP model. Options: asinh, log2, sqrt, none.}
}
\value{
A data frame of input data and predicted TF occupancy (posterior mean).
}
\description{
Predict TF occupancy using posterior mean of TOP regression
coefficients.
}
