% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{predict_TOP_samples}
\alias{predict_TOP_samples}
\title{Predict TF occupancy using posterior samples of regression coefficients}
\usage{
predict_TOP_samples(
  data,
  coef_samples,
  use.posterior.mean = FALSE,
  sample.predictions = TRUE,
  transform = c("asinh", "log2", "sqrt", "none")
)
}
\arguments{
\item{data}{A data frame containing motif PWM score and DNase (or ATAC) bins.}

\item{coef_samples}{TOP posterior samples.}

\item{use.posterior.mean}{Logical; if TRUE, uses the posterior mean of
regression coefficients to make predictions.}

\item{sample.predictions}{Logical; if TRUE, sample from posterior predictions
and then take the mean of posterior prediction samples.}

\item{transform}{Method used to transform ChIP-seq counts when training
the TOP model. Options: asinh, log2, sqrt, none.}
}
\value{
Returns a vector of predicted TF occupancy (posterior mean).
}
\description{
Predict TF occupancy using posterior samples of TOP regression
coefficients.
}
