% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess_Frank.R
\name{PreProcess_Frank}
\alias{PreProcess_Frank}
\title{FUNCTION_TITLE}
\usage{
PreProcess_Frank(
  x_list,
  y_list,
  contrast = NULL,
  nFeatures = 50,
  combinationMethod = "OSP"
)
}
\arguments{
\item{x_list}{PARAM_DESCRIPTION}

\item{y_list}{PARAM_DESCRIPTION}

\item{contrast}{PARAM_DESCRIPTION, Default: NULL}

\item{nFeatures}{PARAM_DESCRIPTION, Default: 50}

\item{combinationMethod}{PARAM_DESCRIPTION, Default: 'OSP'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
