% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_candidate_sites}
\alias{process_candidate_sites}
\title{Process candidate sites from FIMO result}
\usage{
process_candidate_sites(
  file_fimo,
  file_sites,
  flank = 100,
  thresh_pValue = 1e-05,
  compute_mapability = FALSE,
  file_mapability = NA,
  filter_blacklist = FALSE,
  file_blacklist = NA,
  path_bigWigAverageOverBed = "bigWigAverageOverBed",
  path_bedtools = "bedtools"
)
}
\arguments{
\item{file_fimo}{Filename of FIMO result.}

\item{file_sites}{Filename of candidate sites.}

\item{flank}{Size (bp) of flanking region on each side of motif. Default: 100.}

\item{thresh_pValue}{FIMO p-value threshold.}

\item{compute_mapability}{If TRUE, compute mapability for candidate sites.}

\item{file_mapability}{Filename of the mapability reference file in bigWig format.}

\item{filter_blacklist}{If TRUE, filter out sites in blacklist regions.}

\item{file_blacklist}{Filename of the blacklist regions}

\item{path_bigWigAverageOverBed}{Path to bigWigAverageOverBed executable.}

\item{path_bedtools}{Path to bedtools executable.}
}
\description{
Process candidate sites from FIMO result
}
