% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_utils.R
\name{process_candidate_sites}
\alias{process_candidate_sites}
\title{Obtain and filter candidate sites from FIMO result}
\usage{
process_candidate_sites(
  fimo_file,
  flank = 100,
  thresh_pValue = 1e-05,
  thresh_pwmscore = 0,
  blacklist_file,
  mapability_file,
  thresh_mapability = 0.8,
  bigWigAverageOverBed_path = "bigWigAverageOverBed"
)
}
\arguments{
\item{fimo_file}{FIMO result .txt file}

\item{flank}{Flanking region (bp) around motif matches (default: 100)}

\item{thresh_pValue}{FIMO p-value threshold (default: 1e-5)}

\item{thresh_pwmscore}{FIMO PWM score threshold (default: 0)}

\item{blacklist_file}{Filename of the blacklist regions}

\item{mapability_file}{Filename of the mapability reference file in bigWig format.}

\item{thresh_mapability}{Mapability threshold (default: 0.8,
candidate sites need to be mapable at least 80\% positions).}

\item{bigWigAverageOverBed_path}{Path to bigWigAverageOverBed executable,
only needed when filtering mapability.}
}
\description{
Obtain and filter candidate sites from FIMO result
}
