% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase_utils.R
\name{rev_count_bwtool}
\alias{rev_count_bwtool}
\title{Flip the counts generated from bwtool for motifs on the reverse (minus) strand}
\usage{
rev_count_bwtool(sites_file, dnase_fwd_matrix_file, dnase_rev_matrix_file)
}
\arguments{
\item{sites_file}{Filename for candidate sites}

\item{dnase_fwd_matrix_file}{Filename for DNase count matrix in forward strand}

\item{dnase_rev_matrix_file}{Filename for DNase count matrix in reverse strand}
}
\description{
Flip the counts generated from bwtool for motifs on the reverse (minus) strand
}
