% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatterplot}
\alias{scatterplot}
\title{Simple x-y scatter plot}
\usage{
scatterplot(
  x,
  y,
  xlab = "",
  ylab = "",
  title = "",
  xlim = NULL,
  ylim = NULL,
  color = "black"
)
}
\arguments{
\item{x}{x-axis values of points in the plot.}

\item{y}{y-axis values of points in the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{title}{a title for the plot.}

\item{xlim}{range of x-axis values.}

\item{ylim}{range of y-axis values.}

\item{color}{The plotting color.}
}
\description{
Simple x-y scatter plot
}
