% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatterplot_predictions}
\alias{scatterplot_predictions}
\title{Make a scatter plot of measured and predicted occupancy,
with Pearson's correlation (R) between measured and predicted occupancy}
\usage{
scatterplot_predictions(
  x,
  y,
  xlab = "measured",
  ylab = "predicted",
  title = "",
  xlim = c(0, 10),
  ylim = c(0, 10),
  color = "black"
)
}
\arguments{
\item{x}{x-axis values of points in the plot (measured).}

\item{y}{y-axis values of points in the plot (predicted).}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{title}{a title for the plot.}

\item{xlim}{range of x-axis values.}

\item{ylim}{range of y-axis values.}

\item{color}{The plotting color.}
}
\description{
Make a scatter plot of measured and predicted occupancy,
with Pearson's correlation (R) between measured and predicted occupancy
}
