% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTestSupportMixedModel.R
\name{ScoreTestSupportMixedModel}
\alias{ScoreTestSupportMixedModel}
\title{Title}
\usage{
ScoreTestSupportMixedModel(y, baselineonly = NULL, additive = NULL,
  pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888, delta0 = NULL)
}
\arguments{
\item{missingTumorIndicator}{}
}
\description{
Title
}
