% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occupancy.R
\name{select_features}
\alias{select_features}
\title{Select PWM and DNase (or ATAC) bin features from the input data}
\usage{
select_features(data, pwm_col = "pwm", bin_col = "bin")
}
\arguments{
\item{data}{A data frame containing motif PWM score and DNase (or ATAC) bins.}

\item{pwm_col}{name (prefix) of the PMW score column.}

\item{bin_col}{name (prefix) of the DNase or ATAC bin columns.}
}
\description{
Select PWM and DNase (or ATAC) bin features from the input data
}
