% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_model_utils.R
\name{select_model_coef_level}
\alias{select_model_coef_level}
\title{Select regression coefficients by the TOP hierarchy level}
\usage{
select_model_coef_level(
  pwm_id,
  cell_type,
  training_tf_pwm_table,
  TOP_mean_coef,
  level = c("best", "middle", "top")
)
}
\arguments{
\item{pwm_id}{specifies the motif PWM ID.}

\item{cell_type}{specifies the cell type.}

\item{training_tf_pwm_table}{A data frame including TF names and motif PWM IDs
in the training data.}

\item{TOP_mean_coef}{Trained TOP posterior mean regression coefficients.}

\item{level}{Specific the TOP model hierarchy level to use.
Options: 'best', 'middle', or 'top'.
Default: 'best' -- choose the best model (lowest level of the hierarchy available):
If the TF motif and cell type is available in the training data,
then use the bottom level (TF- and cell-type-specific model).
otherwise, if TF motif (but not cell type) is available in the training data,
choose the middle level (TF-specific model) of that TF motif;
otherwise, use the top level TF-generic model.}
}
\description{
Select regression coefficients by the TOP hierarchy level
}
