% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOP_model_utils.R
\name{select_TOP_samples}
\alias{select_TOP_samples}
\title{Select TOP posterior samples}
\usage{
select_TOP_samples(TOP_samples, thin = 1, n.samples = 1000)
}
\arguments{
\item{TOP_samples}{TOP posterior samples.}

\item{thin}{Thinning rate of extract the posterior samples,
must be a positive integer (default = 1, no thinning performed).}

\item{n.samples}{Keep n.samples posterior samples (randomly choose),
when the number of posterior samples is greater than \code{n.samples}.}
}
\value{
A data frame of posterior samples.
}
\description{
Select posterior samples and perform thinning and sampling
of the posterior samples if needed.
}
