% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_utils.R
\name{sort_index_idxstats_bam}
\alias{sort_index_idxstats_bam}
\title{Sort, index the BAM file, and retrieve the idxstats.}
\usage{
sort_index_idxstats_bam(
  bam_file,
  sorted_bam_file,
  sort = TRUE,
  index = TRUE,
  idxstats = TRUE
)
}
\arguments{
\item{bam_file}{Input BAM file.}

\item{sorted_bam_file}{Output file name for sorted BAM file if \code{sort=TRUE}.}

\item{sort}{Logical. If TRUE, sort the BAM file.}

\item{index}{Logical. If TRUE, index the BAM file.}

\item{idxstats}{Logical. If TRUE, retrieve idxstats summary of
the number of mapped reads on every chromosome.}
}
\description{
Sort and index the BAM file, and
retrieve the idxstats (summary of
the number of mapped reads on every chromosome) using \code{Rsamtools}.
}
\examples{
# Sort, index the BAM file, and retrieve the idxstats.
sort_index_idxstats_bam('example.bam', sort=TRUE, index=TRUE, idxstats=TRUE)

# Index the BAM file, and retrieve the idxstats. Assuming the bam file is already sorted.
sort_index_idxstats_bam('example.sorted.bam', sort=FALSE, index=TRUE, idxstats=TRUE)
}
