% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_tracks.R
\name{sort_merge_overlap_bedgraph}
\alias{sort_merge_overlap_bedgraph}
\title{Remove the overlapping site with lower occupancy}
\usage{
sort_merge_overlap_bedgraph(predicted_bedgraph.df)
}
\arguments{
\item{predicted_bedgraph.df}{predicted TF occupancy in bedgraph format}
}
\value{

}
\description{
Remove the overlapping site with lower occupancy
}
