% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_Frank.R
\name{Surv_Frank}
\alias{Surv_Frank}
\title{FUNCTION_TITLE}
\usage{
Surv_Frank(x_list, y_list)
}
\arguments{
\item{x_list}{PARAM_DESCRIPTION}

\item{y_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
