% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{TOP_coefPlot}
\alias{TOP_coefPlot}
\title{TOP_coefPlot}
\usage{
TOP_coefPlot(TOP_model, nFeatures = 20, s = "lambda.min")
}
\arguments{
\item{TOP_model}{A Transferable Omics Prediction model. THe output from the TOP_model function.}

\item{nFeatures}{The number of features that will be plotted. Default: 20}

\item{s}{Lambda value for the lasso model, Default: 'lambda.min'}
}
\description{
TOP_coefPlot
}
\examples{
data(cpop_data_binary, package = "CPOP")

x1 <- cpop_data_binary$x1
x2 <- cpop_data_binary$x2
x3 <- cpop_data_binary$x3
y1 <- cpop_data_binary$y1
y2 <- cpop_data_binary$y2
y3 <- cpop_data_binary$y3

set.seed(23)
x_list <- list(x1, x2)
y_list <- list(factor(y1), factor(y2))

model <- TOP_model(x_list, y_list)
TOP_coefPlot(model)

}
